ChAMP QC Pipeline for beta matrix

ChAMP QC Pipeline for beta matrix

ChAMP_QC_Pipeline_Frome_Beta_Value(
  Sample.beta.df = NULL,
  Sample.Group = "",
  Slide = "",
  arraytype = c("450K", "EPIC"),
  CpG.GUI = FALSE,
  QC.GUI = FALSE,
  combat = FALSE,
  batchname = c("Slide"),
  cores = 50
)

Arguments

Sample.beta.df

Column is sample

arraytype

450K or EPIC

CpG.GUI

If use CpG.GUI

combat

If perform combat to remove batch effect

batchname

Batch variable name. Default is c("Slide")

cores

Examples

library(ChAMP)
#> Loading required package: minfi
#> Loading required package: BiocGenerics
#> 
#> Attaching package: ‘BiocGenerics’
#> The following object is masked from ‘package:loonR’:
#> 
#>     plotPCA
#> The following objects are masked from ‘package:dplyr’:
#> 
#>     combine, intersect, setdiff, union
#> The following object is masked from ‘package:pROC’:
#> 
#>     var
#> The following objects are masked from ‘package:stats’:
#> 
#>     IQR, mad, sd, var, xtabs
#> The following objects are masked from ‘package:base’:
#> 
#>     Filter, Find, Map, Position, Reduce, anyDuplicated, append,
#>     as.data.frame, basename, cbind, colnames, dirname, do.call,
#>     duplicated, eval, evalq, get, grep, grepl, intersect, is.unsorted,
#>     lapply, mapply, match, mget, order, paste, pmax, pmax.int, pmin,
#>     pmin.int, rank, rbind, rownames, sapply, setdiff, sort, table,
#>     tapply, union, unique, unsplit, which.max, which.min
#> Loading required package: GenomicRanges
#> Loading required package: stats4
#> Loading required package: S4Vectors
#> 
#> Attaching package: ‘S4Vectors’
#> The following objects are masked from ‘package:dplyr’:
#> 
#>     first, rename
#> The following objects are masked from ‘package:base’:
#> 
#>     I, expand.grid, unname
#> Loading required package: IRanges
#> 
#> Attaching package: ‘IRanges’
#> The following objects are masked from ‘package:dplyr’:
#> 
#>     collapse, desc, slice
#> Loading required package: GenomeInfoDb
#> Loading required package: SummarizedExperiment
#> Loading required package: MatrixGenerics
#> Loading required package: matrixStats
#> 
#> Attaching package: ‘matrixStats’
#> The following object is masked from ‘package:dplyr’:
#> 
#>     count
#> 
#> Attaching package: ‘MatrixGenerics’
#> The following objects are masked from ‘package:matrixStats’:
#> 
#>     colAlls, colAnyNAs, colAnys, colAvgsPerRowSet, colCollapse,
#>     colCounts, colCummaxs, colCummins, colCumprods, colCumsums,
#>     colDiffs, colIQRDiffs, colIQRs, colLogSumExps, colMadDiffs,
#>     colMads, colMaxs, colMeans2, colMedians, colMins, colOrderStats,
#>     colProds, colQuantiles, colRanges, colRanks, colSdDiffs, colSds,
#>     colSums2, colTabulates, colVarDiffs, colVars, colWeightedMads,
#>     colWeightedMeans, colWeightedMedians, colWeightedSds,
#>     colWeightedVars, rowAlls, rowAnyNAs, rowAnys, rowAvgsPerColSet,
#>     rowCollapse, rowCounts, rowCummaxs, rowCummins, rowCumprods,
#>     rowCumsums, rowDiffs, rowIQRDiffs, rowIQRs, rowLogSumExps,
#>     rowMadDiffs, rowMads, rowMaxs, rowMeans2, rowMedians, rowMins,
#>     rowOrderStats, rowProds, rowQuantiles, rowRanges, rowRanks,
#>     rowSdDiffs, rowSds, rowSums2, rowTabulates, rowVarDiffs, rowVars,
#>     rowWeightedMads, rowWeightedMeans, rowWeightedMedians,
#>     rowWeightedSds, rowWeightedVars
#> Loading required package: Biobase
#> Welcome to Bioconductor
#> 
#>     Vignettes contain introductory material; view with
#>     'browseVignettes()'. To cite Bioconductor, see
#>     'citation("Biobase")', and for packages 'citation("pkgname")'.
#> 
#> Attaching package: ‘Biobase’
#> The following object is masked from ‘package:MatrixGenerics’:
#> 
#>     rowMedians
#> The following objects are masked from ‘package:matrixStats’:
#> 
#>     anyMissing, rowMedians
#> Loading required package: Biostrings
#> Loading required package: XVector
#> 
#> Attaching package: ‘Biostrings’
#> The following object is masked from ‘package:base’:
#> 
#>     strsplit
#> Loading required package: bumphunter
#> Loading required package: foreach
#> Loading required package: iterators
#> Loading required package: parallel
#> Loading required package: locfit
#> locfit 1.5-9.8 	 2023-06-11
#> Setting options('download.file.method.GEOquery'='auto')
#> Setting options('GEOquery.inmemory.gpl'=FALSE)
#> Loading required package: ChAMPdata
#> Loading required package: DMRcate
#> Warning: replacing previous import 'matrixStats::rowMedians' by 'Biobase::rowMedians' when loading 'DSS'
#> Warning: replacing previous import 'matrixStats::anyMissing' by 'Biobase::anyMissing' when loading 'DSS'
#> 
#> Loading required package: Illumina450ProbeVariants.db
#> Loading required package: IlluminaHumanMethylationEPICmanifest
#> Loading required package: DT
#> Loading required package: RPMM
#> Loading required package: cluster
#> No methods found in package 'RSQLite' for request: 'dbListFields' when loading 'lumi'
#> Warning: replacing previous import 'plyr::summarise' by 'plotly::summarise' when loading 'ChAMP'
#> Warning: replacing previous import 'plyr::rename' by 'plotly::rename' when loading 'ChAMP'
#> Warning: replacing previous import 'plyr::arrange' by 'plotly::arrange' when loading 'ChAMP'
#> Warning: replacing previous import 'plyr::mutate' by 'plotly::mutate' when loading 'ChAMP'
#> 
#> Warning: replacing previous import 'plotly::subplot' by 'Hmisc::subplot' when loading 'ChAMP'
#> Warning: replacing previous import 'plyr::is.discrete' by 'Hmisc::is.discrete' when loading 'ChAMP'
#> Warning: replacing previous import 'plyr::summarize' by 'Hmisc::summarize' when loading 'ChAMP'
#> Warning: replacing previous import 'GenomicRanges::sort' by 'globaltest::sort' when loading 'ChAMP'
#> Warning: replacing previous import 'plotly::last_plot' by 'ggplot2::last_plot' when loading 'ChAMP'
#> Warning: replacing previous import 'globaltest::model.matrix' by 'stats::model.matrix' when loading 'ChAMP'
#> Warning: replacing previous import 'globaltest::p.adjust' by 'stats::p.adjust' when loading 'ChAMP'
#> >> Package version 2.21.1 loaded <<
#>        ___ _      _   __  __ ___ 
#>       / __| |_   /_\ |  \/  | _ \
#>      | (__| ' \ / _ \| |\/| |  _/
#>       \___|_||_/_/ \_\_|  |_|_|  
#>       ------------------------------
#>     If you have any question or suggestion about ChAMP, please email to champ450k@gmail.com.
#>     Thank you for citating ChAMP:
#> 
#>     Yuan Tian, Tiffany J Morris, Amy P Webster, Zhen Yang, Stephan Beck, Andrew Feber, Andrew E Teschendorff; ChAMP: updated methylation analysis pipeline for Illumina BeadChips, Bioinformatics, btx513, https://doi.org/10.1093/bioinformatics/btx513
#>      --------------------------
testDir=system.file("extdata",package="ChAMPdata")
myImport <- champ.import(testDir)
#> [===========================]
#> [<<<< ChAMP.IMPORT START >>>>>]
#> -----------------------------
#> 
#> [ Section 1: Read PD Files Start ]
#>   CSV Directory: /master/zhu_zhong_xu/R/x86_64-pc-linux-gnu-library/4.2/ChAMPdata/extdata/lung_test_set.csv
#>   Find CSV Success
#>   Reading CSV File
#> Warning: incomplete final line found on '/master/zhu_zhong_xu/R/x86_64-pc-linux-gnu-library/4.2/ChAMPdata/extdata/lung_test_set.csv'
#>   Replace Sentrix_Position into Array
#>   Replace Sentrix_ID into Slide
#> [ Section 1: Read PD file Done ]
#> 
#> 
#> [ Section 2: Read IDAT files Start ]
#>   Loading:/master/zhu_zhong_xu/R/x86_64-pc-linux-gnu-library/4.2/ChAMPdata/extdata/7990895118_R03C02_Grn.idat ---- (1/8)
#> Warning: truncating string with embedded nuls
#>   Loading:/master/zhu_zhong_xu/R/x86_64-pc-linux-gnu-library/4.2/ChAMPdata/extdata/7990895118_R05C02_Grn.idat ---- (2/8)
#> Warning: truncating string with embedded nuls
#>   Loading:/master/zhu_zhong_xu/R/x86_64-pc-linux-gnu-library/4.2/ChAMPdata/extdata/9247377086_R01C01_Grn.idat ---- (3/8)
#> Warning: truncating string with embedded nuls
#>   Loading:/master/zhu_zhong_xu/R/x86_64-pc-linux-gnu-library/4.2/ChAMPdata/extdata/9247377086_R02C01_Grn.idat ---- (4/8)
#> Warning: truncating string with embedded nuls
#>   Loading:/master/zhu_zhong_xu/R/x86_64-pc-linux-gnu-library/4.2/ChAMPdata/extdata/7766130112_R06C01_Grn.idat ---- (5/8)
#> Warning: truncating string with embedded nuls
#>   Loading:/master/zhu_zhong_xu/R/x86_64-pc-linux-gnu-library/4.2/ChAMPdata/extdata/7766130112_R01C02_Grn.idat ---- (6/8)
#> Warning: truncating string with embedded nuls
#>   Loading:/master/zhu_zhong_xu/R/x86_64-pc-linux-gnu-library/4.2/ChAMPdata/extdata/7990895118_R01C01_Grn.idat ---- (7/8)
#> Warning: truncating string with embedded nuls
#>   Loading:/master/zhu_zhong_xu/R/x86_64-pc-linux-gnu-library/4.2/ChAMPdata/extdata/7990895118_R01C02_Grn.idat ---- (8/8)
#> Warning: truncating string with embedded nuls
#>   Loading:/master/zhu_zhong_xu/R/x86_64-pc-linux-gnu-library/4.2/ChAMPdata/extdata/7990895118_R03C02_Red.idat ---- (1/8)
#> Warning: truncating string with embedded nuls
#>   Loading:/master/zhu_zhong_xu/R/x86_64-pc-linux-gnu-library/4.2/ChAMPdata/extdata/7990895118_R05C02_Red.idat ---- (2/8)
#> Warning: truncating string with embedded nuls
#>   Loading:/master/zhu_zhong_xu/R/x86_64-pc-linux-gnu-library/4.2/ChAMPdata/extdata/9247377086_R01C01_Red.idat ---- (3/8)
#> Warning: truncating string with embedded nuls
#>   Loading:/master/zhu_zhong_xu/R/x86_64-pc-linux-gnu-library/4.2/ChAMPdata/extdata/9247377086_R02C01_Red.idat ---- (4/8)
#> Warning: truncating string with embedded nuls
#>   Loading:/master/zhu_zhong_xu/R/x86_64-pc-linux-gnu-library/4.2/ChAMPdata/extdata/7766130112_R06C01_Red.idat ---- (5/8)
#> Warning: truncating string with embedded nuls
#>   Loading:/master/zhu_zhong_xu/R/x86_64-pc-linux-gnu-library/4.2/ChAMPdata/extdata/7766130112_R01C02_Red.idat ---- (6/8)
#> Warning: truncating string with embedded nuls
#>   Loading:/master/zhu_zhong_xu/R/x86_64-pc-linux-gnu-library/4.2/ChAMPdata/extdata/7990895118_R01C01_Red.idat ---- (7/8)
#> Warning: truncating string with embedded nuls
#>   Loading:/master/zhu_zhong_xu/R/x86_64-pc-linux-gnu-library/4.2/ChAMPdata/extdata/7990895118_R01C02_Red.idat ---- (8/8)
#> Warning: truncating string with embedded nuls
#> 
#>   Extract Mean value for Green and Red Channel Success
#>     Your Red Green Channel contains 622399 probes.
#> [ Section 2: Read IDAT Files Done ]
#> 
#> 
#> [ Section 3: Use Annotation Start ]
#> 
#>   Reading 450K Annotation >>
#> 
#>   Fetching NEGATIVE ControlProbe.
#>     Totally, there are 613 control probes in Annotation.
#>     Your data set contains 613 control probes.
#> 
#>   Generating Meth and UnMeth Matrix
#>     Extracting Meth Matrix...
#>       Totally there are 485512 Meth probes in 450K Annotation.
#>       Your data set contains 485512 Meth probes.
#>     Extracting UnMeth Matrix...
#>       Totally there are 485512 UnMeth probes in 450K Annotation.
#>       Your data set contains 485512 UnMeth probes.
#> 
#>   Generating beta Matrix
#>   Generating M Matrix
#>   Generating intensity Matrix
#>   Calculating Detect P value
#>   Counting Beads
#> [ Section 3: Use Annotation Done ]
#> 
#> [<<<<< ChAMP.IMPORT END >>>>>>]
#> [===========================]
#> [You may want to process champ.filter() next.]

beta.df = myImport$beta
group = myImport$pd$Sample_Group

########### Step 1
res = loonR::ChAMP_QC_Pipeline_Frome_Beta_Value(Sample.beta.df=beta.df, Sample.Group=group)
#> Warning: Start filtering 
#> [===========================]
#> [<<<< ChAMP.FILTER START >>>>>]
#> -----------------------------
#> 
#> In New version ChAMP, champ.filter() function has been set to do filtering on the result of champ.import(). You can use champ.import() + champ.filter() to do Data Loading, or set "method" parameter in champ.load() as "ChAMP" to get the same effect.
#> 
#> This function is provided for user need to do filtering on some beta (or M) matrix, which contained most filtering system in champ.load except beadcount. User need to input beta matrix, pd file themselves. If you want to do filterintg on detP matrix and Bead Count, you also need to input a detected P matrix and Bead Count information.
#> 
#> Note that if you want to filter more data matrix, say beta, M, intensity... please make sure they have exactly the same rownames and colnames.
#> 
#> 
#> [ Section 1:  Check Input Start ]
#>   You have inputed beta for Analysis.
#> 
#>   pd file provided, checking if it's in accord with Data Matrix...
#>     pd file check success.
#> 
#>   Parameter filterDetP is TRUE, checking if detP in accord with Data Matrix...
#>     !!! Parameter detP is not found, filterDetP is reset FALSE now.
#> 
#>   Parameter filterBeads is TRUE, checking if beadcount in accord with Data Matrix...
#>     !!! Parameter beadcount is not found, filterBeads is reset FALSE now.
#> 
#>   parameter autoimpute is TRUE. Checking if the conditions are fulfilled...
#>     !!! ProbeCutoff is 0, which means you have no needs to do imputation. autoimpute has been reset FALSE.
#> 
#>   Checking Finished :filterMultiHit,filterSNPs,filterNoCG,filterXY would be done on beta.
#> [ Section 1: Check Input Done ]
#> 
#> 
#> [ Section 2: Filtering Start >>
#> 
#>   Filtering NoCG Start
#>     Only Keep CpGs, removing 3091 probes from the analysis.
#> 
#>   Filtering SNPs Start
#>     Using general 450K SNP list for filtering.
#>     Filtering probes with SNPs as identified in Zhou's Nucleic Acids Research Paper 2016.
#>     Removing 59901 probes from the analysis.
#> 
#>   Filtering MultiHit Start
#>     Filtering probes that align to multiple locations as identified in Nordlund et al
#>     Removing 11 probes from the analysis.
#> 
#>   Filtering XY Start
#>     Filtering probes located on X,Y chromosome, removing 10028 probes from the analysis.
#> 
#>   Updating PD file
#>     filterDetP parameter is FALSE, so no Sample Would be removed.
#> [ Section 2: Filtering Done ]
#> 
#>  All filterings are Done, now you have 412481 probes and 8 samples.
#> [<<<<< ChAMP.FILTER END >>>>>>]
#> [===========================]
#> [You may want to process champ.QC() next.]
myLoad = res$myLoad

########### Step 2
res = res$followAnalysisImputeQCNormCombat(arraytype="450K")
#> [===========================]
#> [<<< ChAMP.IMPUTE START >>>>]
#> -----------------------------
#> <method>:Combine. (Suitable for Large Data Set)
#> (1): 0 Samples contain 0.2 or above NA will be removed.
#> (2): 0 Probes contain 0.2 or above NA will be removed.
#> Cluster size 412481 broken into 200578 211903 
#> Cluster size 200578 broken into 147638 52940 
#> Cluster size 147638 broken into 44391 103247 
#> Cluster size 44391 broken into 15694 28697 
#> Cluster size 15694 broken into 10967 4727 
#> Cluster size 10967 broken into 6686 4281 
#> Cluster size 6686 broken into 3832 2854 
#> Cluster size 3832 broken into 1620 2212 
#> Cluster size 1620 broken into 1097 523 
#> Done cluster 1097 
#> Done cluster 523 
#> Done cluster 1620 
#> Cluster size 2212 broken into 1611 601 
#> Cluster size 1611 broken into 788 823 
#> Done cluster 788 
#> Done cluster 823 
#> Done cluster 1611 
#> Done cluster 601 
#> Done cluster 2212 
#> Done cluster 3832 
#> Cluster size 2854 broken into 1295 1559 
#> Done cluster 1295 
#> Cluster size 1559 broken into 653 906 
#> Done cluster 653 
#> Done cluster 906 
#> Done cluster 1559 
#> Done cluster 2854 
#> Done cluster 6686 
#> Cluster size 4281 broken into 2130 2151 
#> Cluster size 2130 broken into 1389 741 
#> Done cluster 1389 
#> Done cluster 741 
#> Done cluster 2130 
#> Cluster size 2151 broken into 680 1471 
#> Done cluster 680 
#> Done cluster 1471 
#> Done cluster 2151 
#> Done cluster 4281 
#> Done cluster 10967 
#> Cluster size 4727 broken into 1179 3548 
#> Done cluster 1179 
#> Cluster size 3548 broken into 1675 1873 
#> Cluster size 1675 broken into 1091 584 
#> Done cluster 1091 
#> Done cluster 584 
#> Done cluster 1675 
#> Cluster size 1873 broken into 809 1064 
#> Done cluster 809 
#> Done cluster 1064 
#> Done cluster 1873 
#> Done cluster 3548 
#> Done cluster 4727 
#> Done cluster 15694 
#> Cluster size 28697 broken into 1647 27050 
#> Cluster size 1647 broken into 1156 491 
#> Done cluster 1156 
#> Done cluster 491 
#> Done cluster 1647 
#> Cluster size 27050 broken into 17240 9810 
#> Cluster size 17240 broken into 11479 5761 
#> Cluster size 11479 broken into 3947 7532 
#> Cluster size 3947 broken into 1062 2885 
#> Done cluster 1062 
#> Cluster size 2885 broken into 1420 1465 
#> Done cluster 1420 
#> Done cluster 1465 
#> Done cluster 2885 
#> Done cluster 3947 
#> Cluster size 7532 broken into 3237 4295 
#> Cluster size 3237 broken into 634 2603 
#> Done cluster 634 
#> Cluster size 2603 broken into 966 1637 
#> Done cluster 966 
#> Cluster size 1637 broken into 795 842 
#> Done cluster 795 
#> Done cluster 842 
#> Done cluster 1637 
#> Done cluster 2603 
#> Done cluster 3237 
#> Cluster size 4295 broken into 1981 2314 
#> Cluster size 1981 broken into 833 1148 
#> Done cluster 833 
#> Done cluster 1148 
#> Done cluster 1981 
#> Cluster size 2314 broken into 839 1475 
#> Done cluster 839 
#> Done cluster 1475 
#> Done cluster 2314 
#> Done cluster 4295 
#> Done cluster 7532 
#> Done cluster 11479 
#> Cluster size 5761 broken into 1855 3906 
#> Cluster size 1855 broken into 1587 268 
#> Cluster size 1587 broken into 919 668 
#> Done cluster 919 
#> Done cluster 668 
#> Done cluster 1587 
#> Done cluster 268 
#> Done cluster 1855 
#> Cluster size 3906 broken into 857 3049 
#> Done cluster 857 
#> Cluster size 3049 broken into 1694 1355 
#> Cluster size 1694 broken into 833 861 
#> Done cluster 833 
#> Done cluster 861 
#> Done cluster 1694 
#> Done cluster 1355 
#> Done cluster 3049 
#> Done cluster 3906 
#> Done cluster 5761 
#> Done cluster 17240 
#> Cluster size 9810 broken into 4949 4861 
#> Cluster size 4949 broken into 3165 1784 
#> Cluster size 3165 broken into 2004 1161 
#> Cluster size 2004 broken into 999 1005 
#> Done cluster 999 
#> Done cluster 1005 
#> Done cluster 2004 
#> Done cluster 1161 
#> Done cluster 3165 
#> Cluster size 1784 broken into 1036 748 
#> Done cluster 1036 
#> Done cluster 748 
#> Done cluster 1784 
#> Done cluster 4949 
#> Cluster size 4861 broken into 1597 3264 
#> Cluster size 1597 broken into 896 701 
#> Done cluster 896 
#> Done cluster 701 
#> Done cluster 1597 
#> Cluster size 3264 broken into 1187 2077 
#> Done cluster 1187 
#> Cluster size 2077 broken into 1466 611 
#> Done cluster 1466 
#> Done cluster 611 
#> Done cluster 2077 
#> Done cluster 3264 
#> Done cluster 4861 
#> Done cluster 9810 
#> Done cluster 27050 
#> Done cluster 28697 
#> Done cluster 44391 
#> Cluster size 103247 broken into 52713 50534 
#> Cluster size 52713 broken into 20972 31741 
#> Cluster size 20972 broken into 7005 13967 
#> Cluster size 7005 broken into 4414 2591 
#> Cluster size 4414 broken into 2552 1862 
#> Cluster size 2552 broken into 771 1781 
#> Done cluster 771 
#> Cluster size 1781 broken into 626 1155 
#> Done cluster 626 
#> Done cluster 1155 
#> Done cluster 1781 
#> Done cluster 2552 
#> Cluster size 1862 broken into 734 1128 
#> Done cluster 734 
#> Done cluster 1128 
#> Done cluster 1862 
#> Done cluster 4414 
#> Cluster size 2591 broken into 1790 801 
#> Cluster size 1790 broken into 543 1247 
#> Done cluster 543 
#> Done cluster 1247 
#> Done cluster 1790 
#> Done cluster 801 
#> Done cluster 2591 
#> Done cluster 7005 
#> Cluster size 13967 broken into 11905 2062 
#> Cluster size 11905 broken into 6043 5862 
#> Cluster size 6043 broken into 2891 3152 
#> Cluster size 2891 broken into 1609 1282 
#> Cluster size 1609 broken into 651 958 
#> Done cluster 651 
#> Done cluster 958 
#> Done cluster 1609 
#> Done cluster 1282 
#> Done cluster 2891 
#> Cluster size 3152 broken into 84 3068 
#> Done cluster 84 
#> Cluster size 3068 broken into 1140 1928 
#> Done cluster 1140 
#> Cluster size 1928 broken into 1102 826 
#> Done cluster 1102 
#> Done cluster 826 
#> Done cluster 1928 
#> Done cluster 3068 
#> Done cluster 3152 
#> Done cluster 6043 
#> Cluster size 5862 broken into 4284 1578 
#> Cluster size 4284 broken into 3334 950 
#> Cluster size 3334 broken into 1577 1757 
#> Cluster size 1577 broken into 1119 458 
#> Done cluster 1119 
#> Done cluster 458 
#> Done cluster 1577 
#> Cluster size 1757 broken into 632 1125 
#> Done cluster 632 
#> Done cluster 1125 
#> Done cluster 1757 
#> Done cluster 3334 
#> Done cluster 950 
#> Done cluster 4284 
#> Cluster size 1578 broken into 1251 327 
#> Done cluster 1251 
#> Done cluster 327 
#> Done cluster 1578 
#> Done cluster 5862 
#> Done cluster 11905 
#> Cluster size 2062 broken into 1020 1042 
#> Done cluster 1020 
#> Done cluster 1042 
#> Done cluster 2062 
#> Done cluster 13967 
#> Done cluster 20972 
#> Cluster size 31741 broken into 8510 23231 
#> Cluster size 8510 broken into 3093 5417 
#> Cluster size 3093 broken into 1402 1691 
#> Done cluster 1402 
#> Cluster size 1691 broken into 606 1085 
#> Done cluster 606 
#> Done cluster 1085 
#> Done cluster 1691 
#> Done cluster 3093 
#> Cluster size 5417 broken into 2488 2929 
#> Cluster size 2488 broken into 1689 799 
#> Cluster size 1689 broken into 773 916 
#> Done cluster 773 
#> Done cluster 916 
#> Done cluster 1689 
#> Done cluster 799 
#> Done cluster 2488 
#> Cluster size 2929 broken into 1204 1725 
#> Done cluster 1204 
#> Cluster size 1725 broken into 678 1047 
#> Done cluster 678 
#> Done cluster 1047 
#> Done cluster 1725 
#> Done cluster 2929 
#> Done cluster 5417 
#> Done cluster 8510 
#> Cluster size 23231 broken into 9821 13410 
#> Cluster size 9821 broken into 4568 5253 
#> Cluster size 4568 broken into 747 3821 
#> Done cluster 747 
#> Cluster size 3821 broken into 1789 2032 
#> Cluster size 1789 broken into 1163 626 
#> Done cluster 1163 
#> Done cluster 626 
#> Done cluster 1789 
#> Cluster size 2032 broken into 788 1244 
#> Done cluster 788 
#> Done cluster 1244 
#> Done cluster 2032 
#> Done cluster 3821 
#> Done cluster 4568 
#> Cluster size 5253 broken into 1755 3498 
#> Cluster size 1755 broken into 1138 617 
#> Done cluster 1138 
#> Done cluster 617 
#> Done cluster 1755 
#> Cluster size 3498 broken into 1096 2402 
#> Done cluster 1096 
#> Cluster size 2402 broken into 1093 1309 
#> Done cluster 1093 
#> Done cluster 1309 
#> Done cluster 2402 
#> Done cluster 3498 
#> Done cluster 5253 
#> Done cluster 9821 
#> Cluster size 13410 broken into 8015 5395 
#> Cluster size 8015 broken into 4463 3552 
#> Cluster size 4463 broken into 2589 1874 
#> Cluster size 2589 broken into 1479 1110 
#> Done cluster 1479 
#> Done cluster 1110 
#> Done cluster 2589 
#> Cluster size 1874 broken into 1816 58 
#> Cluster size 1816 broken into 615 1201 
#> Done cluster 615 
#> Done cluster 1201 
#> Done cluster 1816 
#> Done cluster 58 
#> Done cluster 1874 
#> Done cluster 4463 
#> Cluster size 3552 broken into 1467 2085 
#> Done cluster 1467 
#> Cluster size 2085 broken into 1073 1012 
#> Done cluster 1073 
#> Done cluster 1012 
#> Done cluster 2085 
#> Done cluster 3552 
#> Done cluster 8015 
#> Cluster size 5395 broken into 276 5119 
#> Done cluster 276 
#> Cluster size 5119 broken into 2070 3049 
#> Cluster size 2070 broken into 1362 708 
#> Done cluster 1362 
#> Done cluster 708 
#> Done cluster 2070 
#> Cluster size 3049 broken into 732 2317 
#> Done cluster 732 
#> Cluster size 2317 broken into 1206 1111 
#> Done cluster 1206 
#> Done cluster 1111 
#> Done cluster 2317 
#> Done cluster 3049 
#> Done cluster 5119 
#> Done cluster 5395 
#> Done cluster 13410 
#> Done cluster 23231 
#> Done cluster 31741 
#> Done cluster 52713 
#> Cluster size 50534 broken into 25018 25516 
#> Cluster size 25018 broken into 15412 9606 
#> Cluster size 15412 broken into 1551 13861 
#> Cluster size 1551 broken into 535 1016 
#> Done cluster 535 
#> Done cluster 1016 
#> Done cluster 1551 
#> Cluster size 13861 broken into 3557 10304 
#> Cluster size 3557 broken into 2548 1009 
#> Cluster size 2548 broken into 1117 1431 
#> Done cluster 1117 
#> Done cluster 1431 
#> Done cluster 2548 
#> Done cluster 1009 
#> Done cluster 3557 
#> Cluster size 10304 broken into 4659 5645 
#> Cluster size 4659 broken into 1998 2661 
#> Cluster size 1998 broken into 1099 899 
#> Done cluster 1099 
#> Done cluster 899 
#> Done cluster 1998 
#> Cluster size 2661 broken into 1187 1474 
#> Done cluster 1187 
#> Done cluster 1474 
#> Done cluster 2661 
#> Done cluster 4659 
#> Cluster size 5645 broken into 2564 3081 
#> Cluster size 2564 broken into 1622 942 
#> Cluster size 1622 broken into 1200 422 
#> Done cluster 1200 
#> Done cluster 422 
#> Done cluster 1622 
#> Done cluster 942 
#> Done cluster 2564 
#> Cluster size 3081 broken into 1634 1447 
#> Cluster size 1634 broken into 869 765 
#> Done cluster 869 
#> Done cluster 765 
#> Done cluster 1634 
#> Done cluster 1447 
#> Done cluster 3081 
#> Done cluster 5645 
#> Done cluster 10304 
#> Done cluster 13861 
#> Done cluster 15412 
#> Cluster size 9606 broken into 3601 6005 
#> Cluster size 3601 broken into 2600 1001 
#> Cluster size 2600 broken into 1618 982 
#> Cluster size 1618 broken into 555 1063 
#> Done cluster 555 
#> Done cluster 1063 
#> Done cluster 1618 
#> Done cluster 982 
#> Done cluster 2600 
#> Done cluster 1001 
#> Done cluster 3601 
#> Cluster size 6005 broken into 3447 2558 
#> Cluster size 3447 broken into 1509 1938 
#> Cluster size 1509 broken into 659 850 
#> Done cluster 659 
#> Done cluster 850 
#> Done cluster 1509 
#> Cluster size 1938 broken into 944 994 
#> Done cluster 944 
#> Done cluster 994 
#> Done cluster 1938 
#> Done cluster 3447 
#> Cluster size 2558 broken into 1518 1040 
#> Cluster size 1518 broken into 530 988 
#> Done cluster 530 
#> Done cluster 988 
#> Done cluster 1518 
#> Done cluster 1040 
#> Done cluster 2558 
#> Done cluster 6005 
#> Done cluster 9606 
#> Done cluster 25018 
#> Cluster size 25516 broken into 8459 17057 
#> Cluster size 8459 broken into 2517 5942 
#> Cluster size 2517 broken into 1329 1188 
#> Done cluster 1329 
#> Done cluster 1188 
#> Done cluster 2517 
#> Cluster size 5942 broken into 5620 322 
#> Cluster size 5620 broken into 1693 3927 
#> Cluster size 1693 broken into 510 1183 
#> Done cluster 510 
#> Done cluster 1183 
#> Done cluster 1693 
#> Cluster size 3927 broken into 2111 1816 
#> Cluster size 2111 broken into 1450 661 
#> Done cluster 1450 
#> Done cluster 661 
#> Done cluster 2111 
#> Cluster size 1816 broken into 950 866 
#> Done cluster 950 
#> Done cluster 866 
#> Done cluster 1816 
#> Done cluster 3927 
#> Done cluster 5620 
#> Done cluster 322 
#> Done cluster 5942 
#> Done cluster 8459 
#> Cluster size 17057 broken into 6851 10206 
#> Cluster size 6851 broken into 4511 2340 
#> Cluster size 4511 broken into 2538 1973 
#> Cluster size 2538 broken into 1175 1363 
#> Done cluster 1175 
#> Done cluster 1363 
#> Done cluster 2538 
#> Cluster size 1973 broken into 796 1177 
#> Done cluster 796 
#> Done cluster 1177 
#> Done cluster 1973 
#> Done cluster 4511 
#> Cluster size 2340 broken into 952 1388 
#> Done cluster 952 
#> Done cluster 1388 
#> Done cluster 2340 
#> Done cluster 6851 
#> Cluster size 10206 broken into 4907 5299 
#> Cluster size 4907 broken into 2119 2788 
#> Cluster size 2119 broken into 959 1160 
#> Done cluster 959 
#> Done cluster 1160 
#> Done cluster 2119 
#> Cluster size 2788 broken into 1926 862 
#> Cluster size 1926 broken into 1136 790 
#> Done cluster 1136 
#> Done cluster 790 
#> Done cluster 1926 
#> Done cluster 862 
#> Done cluster 2788 
#> Done cluster 4907 
#> Cluster size 5299 broken into 2275 3024 
#> Cluster size 2275 broken into 992 1283 
#> Done cluster 992 
#> Done cluster 1283 
#> Done cluster 2275 
#> Cluster size 3024 broken into 1538 1486 
#> Cluster size 1538 broken into 935 603 
#> Done cluster 935 
#> Done cluster 603 
#> Done cluster 1538 
#> Done cluster 1486 
#> Done cluster 3024 
#> Done cluster 5299 
#> Done cluster 10206 
#> Done cluster 17057 
#> Done cluster 25516 
#> Done cluster 50534 
#> Done cluster 103247 
#> Done cluster 147638 
#> Cluster size 52940 broken into 37548 15392 
#> Cluster size 37548 broken into 19849 17699 
#> Cluster size 19849 broken into 8825 11024 
#> Cluster size 8825 broken into 4456 4369 
#> Cluster size 4456 broken into 1822 2634 
#> Cluster size 1822 broken into 617 1205 
#> Done cluster 617 
#> Done cluster 1205 
#> Done cluster 1822 
#> Cluster size 2634 broken into 1875 759 
#> Cluster size 1875 broken into 113 1762 
#> Done cluster 113 
#> Cluster size 1762 broken into 1111 651 
#> Done cluster 1111 
#> Done cluster 651 
#> Done cluster 1762 
#> Done cluster 1875 
#> Done cluster 759 
#> Done cluster 2634 
#> Done cluster 4456 
#> Cluster size 4369 broken into 2741 1628 
#> Cluster size 2741 broken into 1092 1649 
#> Done cluster 1092 
#> Cluster size 1649 broken into 607 1042 
#> Done cluster 607 
#> Done cluster 1042 
#> Done cluster 1649 
#> Done cluster 2741 
#> Cluster size 1628 broken into 1041 587 
#> Done cluster 1041 
#> Done cluster 587 
#> Done cluster 1628 
#> Done cluster 4369 
#> Done cluster 8825 
#> Cluster size 11024 broken into 8201 2823 
#> Cluster size 8201 broken into 4724 3477 
#> Cluster size 4724 broken into 1397 3327 
#> Done cluster 1397 
#> Cluster size 3327 broken into 1501 1826 
#> Cluster size 1501 broken into 806 695 
#> Done cluster 806 
#> Done cluster 695 
#> Done cluster 1501 
#> Cluster size 1826 broken into 605 1221 
#> Done cluster 605 
#> Done cluster 1221 
#> Done cluster 1826 
#> Done cluster 3327 
#> Done cluster 4724 
#> Cluster size 3477 broken into 1922 1555 
#> Cluster size 1922 broken into 1085 837 
#> Done cluster 1085 
#> Done cluster 837 
#> Done cluster 1922 
#> Cluster size 1555 broken into 1014 541 
#> Done cluster 1014 
#> Done cluster 541 
#> Done cluster 1555 
#> Done cluster 3477 
#> Done cluster 8201 
#> Cluster size 2823 broken into 940 1883 
#> Done cluster 940 
#> Cluster size 1883 broken into 950 933 
#> Done cluster 950 
#> Done cluster 933 
#> Done cluster 1883 
#> Done cluster 2823 
#> Done cluster 11024 
#> Done cluster 19849 
#> Cluster size 17699 broken into 7564 10135 
#> Cluster size 7564 broken into 2749 4815 
#> Cluster size 2749 broken into 1509 1240 
#> Cluster size 1509 broken into 730 779 
#> Done cluster 730 
#> Done cluster 779 
#> Done cluster 1509 
#> Done cluster 1240 
#> Done cluster 2749 
#> Cluster size 4815 broken into 2323 2492 
#> Cluster size 2323 broken into 1316 1007 
#> Done cluster 1316 
#> Done cluster 1007 
#> Done cluster 2323 
#> Cluster size 2492 broken into 1312 1180 
#> Done cluster 1312 
#> Done cluster 1180 
#> Done cluster 2492 
#> Done cluster 4815 
#> Done cluster 7564 
#> Cluster size 10135 broken into 6322 3813 
#> Cluster size 6322 broken into 3064 3258 
#> Cluster size 3064 broken into 1452 1612 
#> Done cluster 1452 
#> Cluster size 1612 broken into 697 915 
#> Done cluster 697 
#> Done cluster 915 
#> Done cluster 1612 
#> Done cluster 3064 
#> Cluster size 3258 broken into 1264 1994 
#> Done cluster 1264 
#> Cluster size 1994 broken into 1077 917 
#> Done cluster 1077 
#> Done cluster 917 
#> Done cluster 1994 
#> Done cluster 3258 
#> Done cluster 6322 
#> Cluster size 3813 broken into 1537 2276 
#> Cluster size 1537 broken into 552 985 
#> Done cluster 552 
#> Done cluster 985 
#> Done cluster 1537 
#> Cluster size 2276 broken into 1271 1005 
#> Done cluster 1271 
#> Done cluster 1005 
#> Done cluster 2276 
#> Done cluster 3813 
#> Done cluster 10135 
#> Done cluster 17699 
#> Done cluster 37548 
#> Cluster size 15392 broken into 6628 8764 
#> Cluster size 6628 broken into 3562 3066 
#> Cluster size 3562 broken into 1783 1779 
#> Cluster size 1783 broken into 1080 703 
#> Done cluster 1080 
#> Done cluster 703 
#> Done cluster 1783 
#> Cluster size 1779 broken into 838 941 
#> Done cluster 838 
#> Done cluster 941 
#> Done cluster 1779 
#> Done cluster 3562 
#> Cluster size 3066 broken into 1505 1561 
#> Cluster size 1505 broken into 862 643 
#> Done cluster 862 
#> Done cluster 643 
#> Done cluster 1505 
#> Cluster size 1561 broken into 933 628 
#> Done cluster 933 
#> Done cluster 628 
#> Done cluster 1561 
#> Done cluster 3066 
#> Done cluster 6628 
#> Cluster size 8764 broken into 3939 4825 
#> Cluster size 3939 broken into 2181 1758 
#> Cluster size 2181 broken into 959 1222 
#> Done cluster 959 
#> Done cluster 1222 
#> Done cluster 2181 
#> Cluster size 1758 broken into 704 1054 
#> Done cluster 704 
#> Done cluster 1054 
#> Done cluster 1758 
#> Done cluster 3939 
#> Cluster size 4825 broken into 2278 2547 
#> Cluster size 2278 broken into 1042 1236 
#> Done cluster 1042 
#> Done cluster 1236 
#> Done cluster 2278 
#> Cluster size 2547 broken into 1256 1291 
#> Done cluster 1256 
#> Done cluster 1291 
#> Done cluster 2547 
#> Done cluster 4825 
#> Done cluster 8764 
#> Done cluster 15392 
#> Done cluster 52940 
#> Done cluster 200578 
#> Cluster size 211903 broken into 133669 78234 
#> Cluster size 133669 broken into 52322 81347 
#> Cluster size 52322 broken into 36395 15927 
#> Cluster size 36395 broken into 16524 19871 
#> Cluster size 16524 broken into 9822 6702 
#> Cluster size 9822 broken into 7650 2172 
#> Cluster size 7650 broken into 3829 3821 
#> Cluster size 3829 broken into 2543 1286 
#> Cluster size 2543 broken into 1163 1380 
#> Done cluster 1163 
#> Done cluster 1380 
#> Done cluster 2543 
#> Done cluster 1286 
#> Done cluster 3829 
#> Cluster size 3821 broken into 2132 1689 
#> Cluster size 2132 broken into 661 1471 
#> Done cluster 661 
#> Done cluster 1471 
#> Done cluster 2132 
#> Cluster size 1689 broken into 947 742 
#> Done cluster 947 
#> Done cluster 742 
#> Done cluster 1689 
#> Done cluster 3821 
#> Done cluster 7650 
#> Cluster size 2172 broken into 1116 1056 
#> Done cluster 1116 
#> Done cluster 1056 
#> Done cluster 2172 
#> Done cluster 9822 
#> Cluster size 6702 broken into 1660 5042 
#> Cluster size 1660 broken into 1022 638 
#> Done cluster 1022 
#> Done cluster 638 
#> Done cluster 1660 
#> Cluster size 5042 broken into 4546 496 
#> Cluster size 4546 broken into 2578 1968 
#> Cluster size 2578 broken into 1088 1490 
#> Done cluster 1088 
#> Done cluster 1490 
#> Done cluster 2578 
#> Cluster size 1968 broken into 858 1110 
#> Done cluster 858 
#> Done cluster 1110 
#> Done cluster 1968 
#> Done cluster 4546 
#> Done cluster 496 
#> Done cluster 5042 
#> Done cluster 6702 
#> Done cluster 16524 
#> Cluster size 19871 broken into 13707 6164 
#> Cluster size 13707 broken into 8834 4873 
#> Cluster size 8834 broken into 4310 4524 
#> Cluster size 4310 broken into 2311 1999 
#> Cluster size 2311 broken into 1139 1172 
#> Done cluster 1139 
#> Done cluster 1172 
#> Done cluster 2311 
#> Cluster size 1999 broken into 1137 862 
#> Done cluster 1137 
#> Done cluster 862 
#> Done cluster 1999 
#> Done cluster 4310 
#> Cluster size 4524 broken into 2076 2448 
#> Cluster size 2076 broken into 867 1209 
#> Done cluster 867 
#> Done cluster 1209 
#> Done cluster 2076 
#> Cluster size 2448 broken into 1168 1280 
#> Done cluster 1168 
#> Done cluster 1280 
#> Done cluster 2448 
#> Done cluster 4524 
#> Done cluster 8834 
#> Cluster size 4873 broken into 2479 2394 
#> Cluster size 2479 broken into 1071 1408 
#> Done cluster 1071 
#> Done cluster 1408 
#> Done cluster 2479 
#> Cluster size 2394 broken into 1327 1067 
#> Done cluster 1327 
#> Done cluster 1067 
#> Done cluster 2394 
#> Done cluster 4873 
#> Done cluster 13707 
#> Cluster size 6164 broken into 3884 2280 
#> Cluster size 3884 broken into 2138 1746 
#> Cluster size 2138 broken into 653 1485 
#> Done cluster 653 
#> Done cluster 1485 
#> Done cluster 2138 
#> Cluster size 1746 broken into 992 754 
#> Done cluster 992 
#> Done cluster 754 
#> Done cluster 1746 
#> Done cluster 3884 
#> Cluster size 2280 broken into 1170 1110 
#> Done cluster 1170 
#> Done cluster 1110 
#> Done cluster 2280 
#> Done cluster 6164 
#> Done cluster 19871 
#> Done cluster 36395 
#> Cluster size 15927 broken into 10432 5495 
#> Cluster size 10432 broken into 3861 6571 
#> Cluster size 3861 broken into 2114 1747 
#> Cluster size 2114 broken into 1533 581 
#> Cluster size 1533 broken into 301 1232 
#> Done cluster 301 
#> Done cluster 1232 
#> Done cluster 1533 
#> Done cluster 581 
#> Done cluster 2114 
#> Cluster size 1747 broken into 752 995 
#> Done cluster 752 
#> Done cluster 995 
#> Done cluster 1747 
#> Done cluster 3861 
#> Cluster size 6571 broken into 2776 3795 
#> Cluster size 2776 broken into 1205 1571 
#> Done cluster 1205 
#> Cluster size 1571 broken into 878 693 
#> Done cluster 878 
#> Done cluster 693 
#> Done cluster 1571 
#> Done cluster 2776 
#> Cluster size 3795 broken into 1503 2292 
#> Cluster size 1503 broken into 258 1245 
#> Done cluster 258 
#> Done cluster 1245 
#> Done cluster 1503 
#> Cluster size 2292 broken into 1058 1234 
#> Done cluster 1058 
#> Done cluster 1234 
#> Done cluster 2292 
#> Done cluster 3795 
#> Done cluster 6571 
#> Done cluster 10432 
#> Cluster size 5495 broken into 3574 1921 
#> Cluster size 3574 broken into 2008 1566 
#> Cluster size 2008 broken into 469 1539 
#> Done cluster 469 
#> Cluster size 1539 broken into 710 829 
#> Done cluster 710 
#> Done cluster 829 
#> Done cluster 1539 
#> Done cluster 2008 
#> Cluster size 1566 broken into 539 1027 
#> Done cluster 539 
#> Done cluster 1027 
#> Done cluster 1566 
#> Done cluster 3574 
#> Cluster size 1921 broken into 1611 310 
#> Cluster size 1611 broken into 1068 543 
#> Done cluster 1068 
#> Done cluster 543 
#> Done cluster 1611 
#> Done cluster 310 
#> Done cluster 1921 
#> Done cluster 5495 
#> Done cluster 15927 
#> Done cluster 52322 
#> Cluster size 81347 broken into 41321 40026 
#> Cluster size 41321 broken into 24688 16633 
#> Cluster size 24688 broken into 8267 16421 
#> Cluster size 8267 broken into 4982 3285 
#> Cluster size 4982 broken into 1557 3425 
#> Cluster size 1557 broken into 878 679 
#> Done cluster 878 
#> Done cluster 679 
#> Done cluster 1557 
#> Cluster size 3425 broken into 1441 1984 
#> Done cluster 1441 
#> Cluster size 1984 broken into 747 1237 
#> Done cluster 747 
#> Done cluster 1237 
#> Done cluster 1984 
#> Done cluster 3425 
#> Done cluster 4982 
#> Cluster size 3285 broken into 1434 1851 
#> Done cluster 1434 
#> Cluster size 1851 broken into 893 958 
#> Done cluster 893 
#> Done cluster 958 
#> Done cluster 1851 
#> Done cluster 3285 
#> Done cluster 8267 
#> Cluster size 16421 broken into 6590 9831 
#> Cluster size 6590 broken into 4664 1926 
#> Cluster size 4664 broken into 2528 2136 
#> Cluster size 2528 broken into 1001 1527 
#> Done cluster 1001 
#> Cluster size 1527 broken into 529 998 
#> Done cluster 529 
#> Done cluster 998 
#> Done cluster 1527 
#> Done cluster 2528 
#> Cluster size 2136 broken into 1027 1109 
#> Done cluster 1027 
#> Done cluster 1109 
#> Done cluster 2136 
#> Done cluster 4664 
#> Cluster size 1926 broken into 1366 560 
#> Done cluster 1366 
#> Done cluster 560 
#> Done cluster 1926 
#> Done cluster 6590 
#> Cluster size 9831 broken into 4009 5822 
#> Cluster size 4009 broken into 1448 2561 
#> Done cluster 1448 
#> Cluster size 2561 broken into 1691 870 
#> Cluster size 1691 broken into 1027 664 
#> Done cluster 1027 
#> Done cluster 664 
#> Done cluster 1691 
#> Done cluster 870 
#> Done cluster 2561 
#> Done cluster 4009 
#> Cluster size 5822 broken into 1023 4799 
#> Done cluster 1023 
#> Cluster size 4799 broken into 2684 2115 
#> Cluster size 2684 broken into 988 1696 
#> Done cluster 988 
#> Cluster size 1696 broken into 992 704 
#> Done cluster 992 
#> Done cluster 704 
#> Done cluster 1696 
#> Done cluster 2684 
#> Cluster size 2115 broken into 770 1345 
#> Done cluster 770 
#> Done cluster 1345 
#> Done cluster 2115 
#> Done cluster 4799 
#> Done cluster 5822 
#> Done cluster 9831 
#> Done cluster 16421 
#> Done cluster 24688 
#> Cluster size 16633 broken into 11285 5348 
#> Cluster size 11285 broken into 9125 2160 
#> Cluster size 9125 broken into 4884 4241 
#> Cluster size 4884 broken into 2936 1948 
#> Cluster size 2936 broken into 2077 859 
#> Cluster size 2077 broken into 1083 994 
#> Done cluster 1083 
#> Done cluster 994 
#> Done cluster 2077 
#> Done cluster 859 
#> Done cluster 2936 
#> Cluster size 1948 broken into 1507 441 
#> Cluster size 1507 broken into 591 916 
#> Done cluster 591 
#> Done cluster 916 
#> Done cluster 1507 
#> Done cluster 441 
#> Done cluster 1948 
#> Done cluster 4884 
#> Cluster size 4241 broken into 1874 2367 
#> Cluster size 1874 broken into 970 904 
#> Done cluster 970 
#> Done cluster 904 
#> Done cluster 1874 
#> Cluster size 2367 broken into 1689 678 
#> Cluster size 1689 broken into 753 936 
#> Done cluster 753 
#> Done cluster 936 
#> Done cluster 1689 
#> Done cluster 678 
#> Done cluster 2367 
#> Done cluster 4241 
#> Done cluster 9125 
#> Cluster size 2160 broken into 473 1687 
#> Done cluster 473 
#> Cluster size 1687 broken into 738 949 
#> Done cluster 738 
#> Done cluster 949 
#> Done cluster 1687 
#> Done cluster 2160 
#> Done cluster 11285 
#> Cluster size 5348 broken into 3378 1970 
#> Cluster size 3378 broken into 668 2710 
#> Done cluster 668 
#> Cluster size 2710 broken into 1053 1657 
#> Done cluster 1053 
#> Cluster size 1657 broken into 745 912 
#> Done cluster 745 
#> Done cluster 912 
#> Done cluster 1657 
#> Done cluster 2710 
#> Done cluster 3378 
#> Cluster size 1970 broken into 1202 768 
#> Done cluster 1202 
#> Done cluster 768 
#> Done cluster 1970 
#> Done cluster 5348 
#> Done cluster 16633 
#> Done cluster 41321 
#> Cluster size 40026 broken into 23436 16590 
#> Cluster size 23436 broken into 16134 7302 
#> Cluster size 16134 broken into 10746 5388 
#> Cluster size 10746 broken into 2379 8367 
#> Cluster size 2379 broken into 834 1545 
#> Done cluster 834 
#> Cluster size 1545 broken into 1201 344 
#> Done cluster 1201 
#> Done cluster 344 
#> Done cluster 1545 
#> Done cluster 2379 
#> Cluster size 8367 broken into 2083 6284 
#> Cluster size 2083 broken into 845 1238 
#> Done cluster 845 
#> Done cluster 1238 
#> Done cluster 2083 
#> Cluster size 6284 broken into 2648 3636 
#> Cluster size 2648 broken into 1490 1158 
#> Done cluster 1490 
#> Done cluster 1158 
#> Done cluster 2648 
#> Cluster size 3636 broken into 1726 1910 
#> Cluster size 1726 broken into 867 859 
#> Done cluster 867 
#> Done cluster 859 
#> Done cluster 1726 
#> Cluster size 1910 broken into 583 1327 
#> Done cluster 583 
#> Done cluster 1327 
#> Done cluster 1910 
#> Done cluster 3636 
#> Done cluster 6284 
#> Done cluster 8367 
#> Done cluster 10746 
#> Cluster size 5388 broken into 1990 3398 
#> Cluster size 1990 broken into 1139 851 
#> Done cluster 1139 
#> Done cluster 851 
#> Done cluster 1990 
#> Cluster size 3398 broken into 1007 2391 
#> Done cluster 1007 
#> Cluster size 2391 broken into 1833 558 
#> Cluster size 1833 broken into 816 1017 
#> Done cluster 816 
#> Done cluster 1017 
#> Done cluster 1833 
#> Done cluster 558 
#> Done cluster 2391 
#> Done cluster 3398 
#> Done cluster 5388 
#> Done cluster 16134 
#> Cluster size 7302 broken into 3388 3914 
#> Cluster size 3388 broken into 1086 2302 
#> Done cluster 1086 
#> Cluster size 2302 broken into 790 1512 
#> Done cluster 790 
#> Cluster size 1512 broken into 862 650 
#> Done cluster 862 
#> Done cluster 650 
#> Done cluster 1512 
#> Done cluster 2302 
#> Done cluster 3388 
#> Cluster size 3914 broken into 1674 2240 
#> Cluster size 1674 broken into 1161 513 
#> Done cluster 1161 
#> Done cluster 513 
#> Done cluster 1674 
#> Cluster size 2240 broken into 1129 1111 
#> Done cluster 1129 
#> Done cluster 1111 
#> Done cluster 2240 
#> Done cluster 3914 
#> Done cluster 7302 
#> Done cluster 23436 
#> Cluster size 16590 broken into 12063 4527 
#> Cluster size 12063 broken into 3735 8328 
#> Cluster size 3735 broken into 2460 1275 
#> Cluster size 2460 broken into 1017 1443 
#> Done cluster 1017 
#> Done cluster 1443 
#> Done cluster 2460 
#> Done cluster 1275 
#> Done cluster 3735 
#> Cluster size 8328 broken into 3210 5118 
#> Cluster size 3210 broken into 1233 1977 
#> Done cluster 1233 
#> Cluster size 1977 broken into 838 1139 
#> Done cluster 838 
#> Done cluster 1139 
#> Done cluster 1977 
#> Done cluster 3210 
#> Cluster size 5118 broken into 2900 2218 
#> Cluster size 2900 broken into 2141 759 
#> Cluster size 2141 broken into 851 1290 
#> Done cluster 851 
#> Done cluster 1290 
#> Done cluster 2141 
#> Done cluster 759 
#> Done cluster 2900 
#> Cluster size 2218 broken into 919 1299 
#> Done cluster 919 
#> Done cluster 1299 
#> Done cluster 2218 
#> Done cluster 5118 
#> Done cluster 8328 
#> Done cluster 12063 
#> Cluster size 4527 broken into 3212 1315 
#> Cluster size 3212 broken into 1487 1725 
#> Done cluster 1487 
#> Cluster size 1725 broken into 1012 713 
#> Done cluster 1012 
#> Done cluster 713 
#> Done cluster 1725 
#> Done cluster 3212 
#> Done cluster 1315 
#> Done cluster 4527 
#> Done cluster 16590 
#> Done cluster 40026 
#> Done cluster 81347 
#> Done cluster 133669 
#> Cluster size 78234 broken into 40545 37689 
#> Cluster size 40545 broken into 18779 21766 
#> Cluster size 18779 broken into 9239 9540 
#> Cluster size 9239 broken into 3672 5567 
#> Cluster size 3672 broken into 2245 1427 
#> Cluster size 2245 broken into 1311 934 
#> Done cluster 1311 
#> Done cluster 934 
#> Done cluster 2245 
#> Done cluster 1427 
#> Done cluster 3672 
#> Cluster size 5567 broken into 2006 3561 
#> Cluster size 2006 broken into 845 1161 
#> Done cluster 845 
#> Done cluster 1161 
#> Done cluster 2006 
#> Cluster size 3561 broken into 2060 1501 
#> Cluster size 2060 broken into 934 1126 
#> Done cluster 934 
#> Done cluster 1126 
#> Done cluster 2060 
#> Cluster size 1501 broken into 796 705 
#> Done cluster 796 
#> Done cluster 705 
#> Done cluster 1501 
#> Done cluster 3561 
#> Done cluster 5567 
#> Done cluster 9239 
#> Cluster size 9540 broken into 3955 5585 
#> Cluster size 3955 broken into 1974 1981 
#> Cluster size 1974 broken into 1209 765 
#> Done cluster 1209 
#> Done cluster 765 
#> Done cluster 1974 
#> Cluster size 1981 broken into 844 1137 
#> Done cluster 844 
#> Done cluster 1137 
#> Done cluster 1981 
#> Done cluster 3955 
#> Cluster size 5585 broken into 2662 2923 
#> Cluster size 2662 broken into 1576 1086 
#> Cluster size 1576 broken into 732 844 
#> Done cluster 732 
#> Done cluster 844 
#> Done cluster 1576 
#> Done cluster 1086 
#> Done cluster 2662 
#> Cluster size 2923 broken into 1697 1226 
#> Cluster size 1697 broken into 792 905 
#> Done cluster 792 
#> Done cluster 905 
#> Done cluster 1697 
#> Done cluster 1226 
#> Done cluster 2923 
#> Done cluster 5585 
#> Done cluster 9540 
#> Done cluster 18779 
#> Cluster size 21766 broken into 12623 9143 
#> Cluster size 12623 broken into 7165 5458 
#> Cluster size 7165 broken into 3002 4163 
#> Cluster size 3002 broken into 1578 1424 
#> Cluster size 1578 broken into 940 638 
#> Done cluster 940 
#> Done cluster 638 
#> Done cluster 1578 
#> Done cluster 1424 
#> Done cluster 3002 
#> Cluster size 4163 broken into 1997 2166 
#> Cluster size 1997 broken into 931 1066 
#> Done cluster 931 
#> Done cluster 1066 
#> Done cluster 1997 
#> Cluster size 2166 broken into 1278 888 
#> Done cluster 1278 
#> Done cluster 888 
#> Done cluster 2166 
#> Done cluster 4163 
#> Done cluster 7165 
#> Cluster size 5458 broken into 3171 2287 
#> Cluster size 3171 broken into 1341 1830 
#> Done cluster 1341 
#> Cluster size 1830 broken into 878 952 
#> Done cluster 878 
#> Done cluster 952 
#> Done cluster 1830 
#> Done cluster 3171 
#> Cluster size 2287 broken into 990 1297 
#> Done cluster 990 
#> Done cluster 1297 
#> Done cluster 2287 
#> Done cluster 5458 
#> Done cluster 12623 
#> Cluster size 9143 broken into 6076 3067 
#> Cluster size 6076 broken into 3655 2421 
#> Cluster size 3655 broken into 1812 1843 
#> Cluster size 1812 broken into 951 861 
#> Done cluster 951 
#> Done cluster 861 
#> Done cluster 1812 
#> Cluster size 1843 broken into 912 931 
#> Done cluster 912 
#> Done cluster 931 
#> Done cluster 1843 
#> Done cluster 3655 
#> Cluster size 2421 broken into 1204 1217 
#> Done cluster 1204 
#> Done cluster 1217 
#> Done cluster 2421 
#> Done cluster 6076 
#> Cluster size 3067 broken into 1967 1100 
#> Cluster size 1967 broken into 1138 829 
#> Done cluster 1138 
#> Done cluster 829 
#> Done cluster 1967 
#> Done cluster 1100 
#> Done cluster 3067 
#> Done cluster 9143 
#> Done cluster 21766 
#> Done cluster 40545 
#> Cluster size 37689 broken into 21572 16117 
#> Cluster size 21572 broken into 6713 14859 
#> Cluster size 6713 broken into 2295 4418 
#> Cluster size 2295 broken into 1324 971 
#> Done cluster 1324 
#> Done cluster 971 
#> Done cluster 2295 
#> Cluster size 4418 broken into 2984 1434 
#> Cluster size 2984 broken into 2053 931 
#> Cluster size 2053 broken into 1204 849 
#> Done cluster 1204 
#> Done cluster 849 
#> Done cluster 2053 
#> Done cluster 931 
#> Done cluster 2984 
#> Done cluster 1434 
#> Done cluster 4418 
#> Done cluster 6713 
#> Cluster size 14859 broken into 4909 9950 
#> Cluster size 4909 broken into 1700 3209 
#> Cluster size 1700 broken into 837 863 
#> Done cluster 837 
#> Done cluster 863 
#> Done cluster 1700 
#> Cluster size 3209 broken into 1382 1827 
#> Done cluster 1382 
#> Cluster size 1827 broken into 831 996 
#> Done cluster 831 
#> Done cluster 996 
#> Done cluster 1827 
#> Done cluster 3209 
#> Done cluster 4909 
#> Cluster size 9950 broken into 4837 5113 
#> Cluster size 4837 broken into 2294 2543 
#> Cluster size 2294 broken into 1009 1285 
#> Done cluster 1009 
#> Done cluster 1285 
#> Done cluster 2294 
#> Cluster size 2543 broken into 1126 1417 
#> Done cluster 1126 
#> Done cluster 1417 
#> Done cluster 2543 
#> Done cluster 4837 
#> Cluster size 5113 broken into 2573 2540 
#> Cluster size 2573 broken into 1515 1058 
#> Cluster size 1515 broken into 636 879 
#> Done cluster 636 
#> Done cluster 879 
#> Done cluster 1515 
#> Done cluster 1058 
#> Done cluster 2573 
#> Cluster size 2540 broken into 1118 1422 
#> Done cluster 1118 
#> Done cluster 1422 
#> Done cluster 2540 
#> Done cluster 5113 
#> Done cluster 9950 
#> Done cluster 14859 
#> Done cluster 21572 
#> Cluster size 16117 broken into 6831 9286 
#> Cluster size 6831 broken into 3227 3604 
#> Cluster size 3227 broken into 1908 1319 
#> Cluster size 1908 broken into 829 1079 
#> Done cluster 829 
#> Done cluster 1079 
#> Done cluster 1908 
#> Done cluster 1319 
#> Done cluster 3227 
#> Cluster size 3604 broken into 1598 2006 
#> Cluster size 1598 broken into 862 736 
#> Done cluster 862 
#> Done cluster 736 
#> Done cluster 1598 
#> Cluster size 2006 broken into 943 1063 
#> Done cluster 943 
#> Done cluster 1063 
#> Done cluster 2006 
#> Done cluster 3604 
#> Done cluster 6831 
#> Cluster size 9286 broken into 5324 3962 
#> Cluster size 5324 broken into 2873 2451 
#> Cluster size 2873 broken into 1641 1232 
#> Cluster size 1641 broken into 724 917 
#> Done cluster 724 
#> Done cluster 917 
#> Done cluster 1641 
#> Done cluster 1232 
#> Done cluster 2873 
#> Cluster size 2451 broken into 1420 1031 
#> Done cluster 1420 
#> Done cluster 1031 
#> Done cluster 2451 
#> Done cluster 5324 
#> Cluster size 3962 broken into 1194 2768 
#> Done cluster 1194 
#> Cluster size 2768 broken into 1249 1519 
#> Done cluster 1249 
#> Cluster size 1519 broken into 650 869 
#> Done cluster 650 
#> Done cluster 869 
#> Done cluster 1519 
#> Done cluster 2768 
#> Done cluster 3962 
#> Done cluster 9286 
#> Done cluster 16117 
#> Done cluster 37689 
#> Done cluster 78234 
#> Done cluster 211903 
#> (3): The rest NA are imputed by KNN method which parameter k as 5.
#> [<<<<< ChAMP.IMPUTE END >>>>]
#> [===========================]
#> Warning: Start imputation 
#> Warning: Start QC 
#> [===========================]
#> [<<<<< ChAMP.QC START >>>>>>]
#> -----------------------------
#> champ.QC Results will be saved in 1stQC
#> [QC plots will be proceed with 412481 probes and 8 samples.]
#> [0 Zeros dectect in your dataset, will be replaced with 0.000001]
#> Error in champ.QC(beta = myLoad$beta, resultsDir = "1stQC"): object 'myLoad' not found
head(res$myNorm)
#> NULL