Forest plot

Forest plot

forest_plot(
  tabletext,
  estimate.data,
  appendHeader = NULL,
  specify.summary = NULL,
  clipping = c(0.1, 4),
  graph.pos = "right",
  xlab = "",
  xlog = TRUE,
  xticks = c(0.1, 0.5, seq(1, 4, 1)),
  x.font.size = 1
)

Arguments

tabletext

data.frame. Column names will be used as header.

estimate.data

Estimate, upper CI, lower CI. Corresponded with tabletext

appendHeader

If provide a vector, top header will be added

specify.summary

Specify the row index. Row will be bold character.

clipping

Range to cut. Will use arrow at two cutting points

graph.pos

Column index, or "left" or "right". Column index to put in right section

xlab

xlab

xlog

If in log-format. If TRUE, vectical line in 0, otherwise a vectical line at 0

xticks

Default c(0.1, 0.5, 1, 2, 3, 4)

x.font.size

x-axis label font size.

Examples

data(LIRI)
or.res <- loonR::univariate_or(LIRI[,c(1,3,4)],LIRI$status)
#> Waiting for profiling to be done...
#> Waiting for profiling to be done...
#> Waiting for profiling to be done...

estimate.data = or.res[,c(2,3,4)]
text.data = data.frame(Variate = or.res$Variate, OR = or.res$OR)

loonR::forest_plot(text.data, estimate.data, graph.pos = 2, specify.summary = 1)
#> Loading required package: forestplot
#> Warning: there is no package called 'forestplot'
#> 'getOption("repos")' replaces Bioconductor standard repositories, see
#> 'help("repositories", package = "BiocManager")' for details.
#> Replacement repositories:
#>     CRAN: https://cran.rstudio.com/
#> Bioconductor version 3.15 (BiocManager 1.30.22), R 4.2.1 (2022-06-23)
#> Installing package(s) 'forestplot'
#> Installation paths not writeable, unable to update packages
#>   path: /opt/R/4.2.1/lib/R/library
#>   packages:
#>     KernSmooth, MASS, Matrix, boot, brew, cachem, callr, class, cli, cluster,
#>     codetools, commonmark, cpp11, crayon, curl, digest, downlit, evaluate,
#>     fansi, fastmap, fontawesome, foreign, fs, gert, gh, highr, htmltools,
#>     htmlwidgets, httr, jsonlite, knitr, lattice, lifecycle, mgcv, nlme, nnet,
#>     openssl, pillar, pkgbuild, pkgload, processx, profvis, ps, purrr, remotes,
#>     rlang, rpart, rstudioapi, sourcetools, spatial, stringi, stringr, survival,
#>     sys, tibble, tinytex, usethis, utf8, vctrs, waldo, whisker, xfun, yaml, zip
#> Old packages: 'RCurl', 'XML', 'colourpicker', 'glmnet', 'gtable', 'igraph',
#>   'markdown', 'mvtnorm', 'recipes', 'wk', 'xml2', 'yulab.utils'
#> Error in forestplot(tabletext, estimate.data, hrzl_lines = hrzl_lines,     new_page = TRUE, vertices = TRUE, is.summary = summary.label,     graph.pos = graph.pos, clip = clipping, xlog = TRUE, boxsize = 0.15,     xticks = xticks, col = fpColors(box = "black", line = "black",         summary = "black", hrz_lines = "black"), xlab = xlab,     txt_gp = fpTxtGp(ticks = gpar(cex = x.font.size))): could not find function "forestplot"