Nomogram plot by rms

Nomogram plot by rms

nomogram.plot(fit = NULL, data = NULL, fun.list = NA, lp = F)

Arguments

fit
rms model

data
The data used to build the model

fun.list
an optional function to transform the linear predictors, and to plot on another axis. If more than one transformation is plotted, put them in a list, e.g. list(function(x) x/2, function(x) 2*x). Any function values equal to NA will be ignored.

lp
If fun.list is NA, lp will be TRUE. Set to FALSE to suppress creation of an axis for scoring X beta

Value

Examples

# Logistic model
data(LIRI)
res = loonR::build.logistic.model(LIRI[,3:5], LIRI$status, rms = T, scale = F)
#> Pls note: Second unique variable is defined as experiment group
#> Setting levels: control = 0, case = 1
#> Setting direction: controls < cases
f1 = loonR::logit2prob
nomogram.plot(res$model, res$data, lp = T)
#> Error in value.chk(at, i, NA, -nint, Limval, type.range = "full") variable ANLN does not have limits defined by datadist
nomogram.plot(res$model, res$data, f1, lp =F)
#> Error in value.chk(at, i, NA, -nint, Limval, type.range = "full") variable ANLN does not have limits defined by datadist

# Survial model
res = build.psm.regression.model(LIRI[,3:5],LIRI$status, LIRI$time, scale = F)
#> Error in Design(m, formula = formula, specials = c("strata", "cluster")) dataset ddist not found for options(datadist=)

surv <- Survival(res$model) # This would also work if f was from cph
#> Error in UseMethod("Survival") no applicable method for 'Survival' applied to an object of class "c('lrm', 'rms', 'glm')"
surv_100 <- function(x) surv(100, lp = x)
surv_300 <- function(x) surv(300, lp = x)

med <- Quantile(res$model)
#> Error in UseMethod("Quantile") no applicable method for 'Quantile' applied to an object of class "c('lrm', 'rms', 'glm')"
med_f <- function(x) med(lp=x)

f.list=list(
  `Median Survival Time`= med_f,
  `Probability of 100-day Survival`=surv_100,
  `Probability of 300-day Survival`=surv_300
)
nomogram.plot(res$model, res$data, fun.list = f.list, lp =F)
#> Error in value.chk(at, i, NA, -nint, Limval, type.range = "full") variable ANLN does not have limits defined by datadist