Plotting the Net benefit function

Plotting the Net benefit function

# S3 method for ntbft
plot(
  nb,
  nolines = 2:dim(nb)[2],
  nobands = NULL,
  ymin = -0.1,
  ymax = max(nb[, c(nolines, nobands)], na.rm = T),
  legpos = c(0.9, 0.8),
  palette = "aaas"
)

Arguments

nb
Object from

nolines
the number of the columns of nb which should be plotted using lines. The default is to plot all columns (except the first one containing the threshold).

nobands
the number of the columns of nb which should be plotted using bands (useful to plot confidence intervals). The default is to plot no bands

legpos
a vector of two coordinates indicating where the legend should be in the graph

palette

Value

Examples

data(LIRI)

pred <- as.vector( unlist( LIRI[,c(3)] ) )
ntbft.boot(LIRI$status, pred)
#> .....................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................
#> 
#> ORDINARY NONPARAMETRIC BOOTSTRAP
#> 
#> 
#> Call:
#> boot(data = pred, statistic = diffnet, R = boots, outcome = outcome)
#> 
#> 
#> Bootstrap Statistics :
#>      original        bias    std. error
#> t1*         0 -0.0012713340  0.02601793
#> t2*         0 -0.0012843068  0.02628342
#> t3*         0 -0.0012996801  0.02653991
#> t4*         0 -0.0013139368  0.02681162
#> t5*         0 -0.0013284936  0.02708912
#> t6*         0 -0.0013433602  0.02737260
#> t7*         0 -0.0013585465  0.02766224
#> t8*         0 -0.0013740630  0.02795824
#> t9*         0 -0.0013899204  0.02826083
#> t10*        0 -0.0014061303  0.02857020
#> t11*        0 -0.0014227044  0.02888660
#> t12*        0 -0.0014396552  0.02921027
#> t13*        0 -0.0014569956  0.02954144
#> t14*        0 -0.0014747394  0.02988040
#> t15*        0 -0.0014929006  0.03022740
#> t16*        0 -0.0015114943  0.03058275
#> t17*        0 -0.0015270046  0.03092101
#> t18*        0 -0.0013652229  0.03104914
#> t19*        0 -0.0013849510  0.03142710
#> t20*        0 -0.0014051724  0.03181473
#> t21*        0 -0.0014144478  0.03220073
#> t22*        0 -0.0014350133  0.03260911
#> t23*        0 -0.0014561129  0.03302841
#> t24*        0 -0.0013797641  0.03340320
#> t25*        0 -0.0013965517  0.03384385
#> t26*        0 -0.0014137931  0.03429685
#> t27*        0 -0.0014857109  0.03478683
#> t28*        0 -0.0015067050  0.03526849
#> t29*        0 -0.0014754735  0.03587198
#> t30*        0 -0.0014950739  0.03639111
#> t31*        0 -0.0016353073  0.03689128
#> t32*        0 -0.0016617647  0.03744348
#> t33*        0 -0.0017059959  0.03808113
#> t34*        0 -0.0017348485  0.03867372
#> t35*        0 -0.0017970822  0.03912911
#> t36*        0 -0.0018292026  0.03975649
#> t37*        0 -0.0018623426  0.04040502
#> t38*        0 -0.0018648498  0.04108026
#> t39*        0 -0.0018051159  0.04177206
#> t40*        0 -0.0017442529  0.04245700
#> t41*        0 -0.0017725015  0.04320685
#> t42*        0 -0.0018017241  0.04398423
#> t43*        0 -0.0018319722  0.04496831
#> t44*        0 -0.0017955665  0.04606810
#> t45*        0 -0.0018252351  0.04695941
#> t46*        0 -0.0018560026  0.04788567
#> t47*        0 -0.0018879310  0.04884885
#> t48*        0 -0.0021518568  0.04979467
#> t49*        0 -0.0021957404  0.05083922
#> t50*        0 -0.0022413793  0.05192774
#> t51*        0 -0.0022888811  0.05306292
#> t52*        0 -0.0023103448  0.05448384
#> t53*        0 -0.0023607850  0.05573441
#> t54*        0 -0.0020086207  0.05724145
#> t55*        0 -0.0016465517  0.05886348
#> t56*        0 -0.0009678683  0.06002253
#> t57*        0 -0.0008718925  0.06199349
#> t58*        0 -0.0008633005  0.06361079
#> t59*        0 -0.0007674516  0.06585350
#> t60*        0 -0.0011939655  0.06792869
#> t61*        0 -0.0011989390  0.06985504
#> t62*        0 -0.0012041742  0.07188624
#> t63*        0 -0.0012096925  0.07403076
#> t64*        0 -0.0012155172  0.07629801
#> t65*        0 -0.0012216749  0.07872109
#> t66*        0 -0.0004584178  0.08168408
#> t67*        0 -0.0004299896  0.08443210
#> t68*        0 -0.0003997845  0.08735636
#> t69*        0  0.0002463849  0.09013311
#> t70*        0  0.0013764368  0.09438725
#> t71*        0  0.0010963139  0.09874891
#> t72*        0  0.0017955665  0.10293673
#> t73*        0  0.0019591315  0.10773842
#> t74*        0  0.0021114058  0.11234011
#> t75*        0  0.0022758621  0.11731574
#> t76*        0  0.0024540230  0.12271199
#> t77*        0  0.0026476762  0.12858364
#> t78*        0  0.0028589342  0.13499545
#> t79*        0  0.0030903120  0.14202449
#> t80*        0  0.0033448276  0.14976326
#> t81*        0  0.0036261343  0.15832373
#> t82*        0  0.0027605364  0.16875179
#> t83*        0  0.0024934077  0.17784138
#> t84*        0  0.0026659483  0.19004070
#> t85*        0  0.0029568966  0.20367004
#> t86*        0  0.0044544335  0.22030557
#> t87*        0  0.0049423077  0.23841059
#> t88*        0  0.0055114943  0.25954404
#> t89*        0  0.0085556426  0.28523447
#> t90*        0  0.0086982759  0.31387263
#> t91*        0  0.0137308429  0.34985884
#> t92*        0  0.0147629310  0.39780583
#> t93*        0  0.0133165025  0.45772378
#> t94*        0  0.0158821839  0.53679281
#> t95*        0  0.0194741379  0.64751440
#> t96*        0  0.0248620690  0.81362748
#> t97*        0  0.0380229885  1.08642420
#> t98*        0  0.0433534483  1.64621817
#> t99*        0  0.0886120690  3.31015737
plot.ntbft(ntbft.boot)
#> Error in plot.ntbft(ntbft.boot) could not find function "plot.ntbft"