Calibration plot

Calibration plot

riskCalibrationPlotSingleComplex(risk, label, color = "#4682B4", bins = 10)

Arguments

risk
Should be risk probability

bins
Default 10

Value

Examples

data("LIRI")
risk = LIRI$ANLN
risk = loonR::logit2prob(scale(risk))
label = LIRI$status
loonR::riskCalibrationPlotSingleComplex(risk, label)
#> Loading required package: CalibrationCurves
#> Downloading GitHub repo BavoDC/CalibrationCurves@HEAD
#> utf8         (1.2.1     -> 1.2.2    ) [CRAN]
#> cli          (3.2.0     -> 3.4.0    ) [CRAN]
#> colorspace   (2.0-1     -> 2.0-3    ) [CRAN]
#> pillar       (1.7.0     -> 1.8.1    ) [CRAN]
#> magrittr     (2.0.2     -> 2.0.3    ) [CRAN]
#> fansi        (0.4.2     -> 1.0.3    ) [CRAN]
#> RColorBrewer (1.1-2     -> 1.1-3    ) [CRAN]
#> R6           (2.5.0     -> 2.5.1    ) [CRAN]
#> lifecycle    (1.0.1     -> 1.0.2    ) [CRAN]
#> farver       (2.1.0     -> 2.1.1    ) [CRAN]
#> withr        (2.4.2     -> 2.5.0    ) [CRAN]
#> tibble       (3.1.6     -> 3.1.8    ) [CRAN]
#> scales       (1.1.1     -> 1.2.1    ) [CRAN]
#> rlang        (1.0.2     -> 1.0.5    ) [CRAN]
#> isoband      (0.2.4     -> 0.2.5    ) [CRAN]
#> gtable       (0.3.0     -> 0.3.1    ) [CRAN]
#> digest       (0.6.27    -> 0.6.29   ) [CRAN]
#> viridisLite  (0.4.0     -> 0.4.1    ) [CRAN]
#> RcppEigen    (0.3.3.9.1 -> 0.3.3.9.2) [CRAN]
#> deldir       (0.2-10    -> 1.0-6    ) [CRAN]
#> Rcpp         (1.0.6     -> 1.0.9    ) [CRAN]
#> interp       (NA        -> 1.1-3    ) [CRAN]
#> jpeg         (0.1-8.1   -> 0.1-9    ) [CRAN]
#> stringi      (1.5.3     -> 1.7.8    ) [CRAN]
#> xfun         (0.30      -> 0.33     ) [CRAN]
#> evaluate     (0.14      -> 0.16     ) [CRAN]
#> yaml         (2.2.1     -> 2.3.5    ) [CRAN]
#> jsonlite     (1.7.2     -> 1.8.0    ) [CRAN]
#> backports    (1.2.1     -> 1.4.1    ) [CRAN]
#> rstudioapi   (0.13      -> 0.14     ) [CRAN]
#> htmltools    (0.5.2     -> 0.5.3    ) [CRAN]
#> htmlwidgets  (1.5.3     -> 1.5.4    ) [CRAN]
#> checkmate    (2.0.0     -> 2.1.0    ) [CRAN]
#> knitr        (1.37      -> 1.40     ) [CRAN]
#> stringr      (1.4.0     -> 1.4.1    ) [CRAN]
#> zoo          (1.8-9     -> 1.8-11   ) [CRAN]
#> sandwich     (3.0-0     -> 3.0-2    ) [CRAN]
#> TH.data      (1.0-10    -> 1.1-1    ) [CRAN]
#> mvtnorm      (1.1-1     -> 1.1-3    ) [CRAN]
#> MatrixModels (0.5-0     -> 0.5-1    ) [CRAN]
#> viridis      (0.6.0     -> 0.6.2    ) [CRAN]
#> htmlTable    (2.1.0     -> 2.4.1    ) [CRAN]
#> data.table   (1.14.0    -> 1.14.2   ) [CRAN]
#> latticeExtra (0.6-29    -> 0.6-30   ) [CRAN]
#> multcomp     (1.4-17    -> 1.4-20   ) [CRAN]
#> polspline    (1.1.19    -> 1.1.20   ) [CRAN]
#> quantreg     (5.85      -> 5.94     ) [CRAN]
#> Hmisc        (4.5-0     -> 4.7-1    ) [CRAN]
#> rms          (6.2-0     -> 6.3-0    ) [CRAN]
#> Installing 49 packages: utf8, cli, colorspace, pillar, magrittr, fansi, RColorBrewer, R6, lifecycle, farver, withr, tibble, scales, rlang, isoband, gtable, digest, viridisLite, RcppEigen, deldir, Rcpp, interp, jpeg, stringi, xfun, evaluate, yaml, jsonlite, backports, rstudioapi, htmltools, htmlwidgets, checkmate, knitr, stringr, zoo, sandwich, TH.data, mvtnorm, MatrixModels, viridis, htmlTable, data.table, latticeExtra, multcomp, polspline, quantreg, Hmisc, rms
#> Installing packages into '/tmp/RtmpjxnD3Y/temp_libpath298a147ad2853'
#> (as 'lib' is unspecified)
#> also installing the dependency 'Matrix'
#>      checking for file ‘/tmp/RtmpGKqfgW/remotes29cc61c7b35e2/BavoDC-CalibrationCurves-2ead53c/DESCRIPTION’ ...  ✓  checking for file ‘/tmp/RtmpGKqfgW/remotes29cc61c7b35e2/BavoDC-CalibrationCurves-2ead53c/DESCRIPTION’ (543ms)
#>   ─  preparing ‘CalibrationCurves’:
#>    checking DESCRIPTION meta-information ...  ✓  checking DESCRIPTION meta-information
#>   ─  checking for LF line-endings in source and make files and shell scripts
#>   ─  checking for empty or unneeded directories
#> ─  building ‘CalibrationCurves_0.1.5.tar.gz’
#>      
#> 
#> Installing package into '/tmp/RtmpjxnD3Y/temp_libpath298a147ad2853'
#> (as 'lib' is unspecified)
#> Loading required package: CalibrationCurves
#> Error in Design(data, formula = formula, specials = "strat") dataset dd not found for options(datadist=)