Split sample by group into traning and validation set

Split sample by group into traning and validation set

splitSampleByGroup(sample = NULL, group = NULL, seed = 666, fraction = 0.5)

Arguments

sample
Vector

group
Vector

seed
Default 666

fraction
Default 0.5

Value

list(Train=train.design, Validation=validation.design)