Combine dpis, psi and tpm file together I modified the code from https://github.com/comprna/SUPPA/blob/master/scripts/Volcano_MA_plot.R

Combine dpis, psi and tpm file together I modified the code from https://github.com/comprna/SUPPA/blob/master/scripts/Volcano_MA_plot.R

suppa.get.final.table(
  sample.names = "",
  psi = "",
  dpsi = "",
  event.tpm = "",
  pval.cutoff = 0.05,
  dpsi.cutoff = 0.3,
  tpm.cutoff = 0
)

Arguments

sample.names

A vector specify the column name of each psi in psivec file.

psi

Psivec file. Generate by Suppa2 diffSplice

dpsi

Dpsi file. Generate by Suppa2 diffSplice file

event.tpm

Event TPM file. Should be specified in Suppa analysis by --save_tpm_events

pval.cutoff

P-value cutoff

dpsi.cutoff

ΔPSI cutoff

tpm.cutoff

Mean TPM cutoff

Value

A data.frame

Examples

suppa.get.final.table(sample.names = smp.names,
         dpsi = "analysis/05suppa/diffSplice.events.dpsi.temp.0",
         psi  = "analysis/05suppa/diffSplice.events.psivec",
         event.tpm = "analysis/05suppa/diffSplice.events_avglogtpm.tab",
         dpsi.cutoff = dpsi.cutoff, pval.cutoff = pval.cutoff, tpm.cutoff = tpm.cutoff)
#> Error in suppa.get.final.table(sample.names = smp.names, dpsi = "analysis/05suppa/diffSplice.events.dpsi.temp.0",     psi = "analysis/05suppa/diffSplice.events.psivec", event.tpm = "analysis/05suppa/diffSplice.events_avglogtpm.tab",     dpsi.cutoff = dpsi.cutoff, pval.cutoff = pval.cutoff, tpm.cutoff = tpm.cutoff): object 'smp.names' not found