ChAMP QC Pipeline for beta matrix

ChAMP QC Pipeline for beta matrix

ChAMP_QC_Pipeline_Frome_Beta_Value(
  Sample.beta.df = NULL,
  Sample.Group = "",
  Slide = "",
  arraytype = c("450K", "EPIC"),
  CpG.GUI = FALSE,
  QC.GUI = FALSE,
  combat = FALSE,
  batchname = c("Slide")
)

Arguments

Sample.beta.df
Column is sample

arraytype
450K or EPIC

CpG.GUI
If use CpG.GUI

combat
If perform combat to remove batch effect

batchname
Batch variable name. Default is c("Slide")

Value

Examples

library(ChAMP)
#> Loading required package: minfi
#> Loading required package: BiocGenerics
#> Loading required package: parallel
#> 
#> Attaching package: ‘BiocGenerics’
#> The following objects are masked from ‘package:parallel’:
#> 
#>     clusterApply, clusterApplyLB, clusterCall, clusterEvalQ,
#>     clusterExport, clusterMap, parApply, parCapply, parLapply,
#>     parLapplyLB, parRapply, parSapply, parSapplyLB
#> The following object is masked from ‘package:loonR’:
#> 
#>     plotPCA
#> The following objects are masked from ‘package:dplyr’:
#> 
#>     combine, intersect, setdiff, union
#> The following object is masked from ‘package:pROC’:
#> 
#>     var
#> The following objects are masked from ‘package:stats’:
#> 
#>     IQR, mad, sd, var, xtabs
#> The following objects are masked from ‘package:base’:
#> 
#>     anyDuplicated, append, as.data.frame, basename, cbind, colnames,
#>     dirname, do.call, duplicated, eval, evalq, Filter, Find, get, grep,
#>     grepl, intersect, is.unsorted, lapply, Map, mapply, match, mget,
#>     order, paste, pmax, pmax.int, pmin, pmin.int, Position, rank,
#>     rbind, Reduce, rownames, sapply, setdiff, sort, table, tapply,
#>     union, unique, unsplit, which.max, which.min
#> Loading required package: GenomicRanges
#> Loading required package: stats4
#> Loading required package: S4Vectors
#> 
#> Attaching package: ‘S4Vectors’
#> The following objects are masked from ‘package:dplyr’:
#> 
#>     first, rename
#> The following object is masked from ‘package:base’:
#> 
#>     expand.grid
#> Loading required package: IRanges
#> 
#> Attaching package: ‘IRanges’
#> The following objects are masked from ‘package:dplyr’:
#> 
#>     collapse, desc, slice
#> Loading required package: GenomeInfoDb
#> Loading required package: SummarizedExperiment
#> Loading required package: MatrixGenerics
#> Loading required package: matrixStats
#> 
#> Attaching package: ‘matrixStats’
#> The following object is masked from ‘package:dplyr’:
#> 
#>     count
#> 
#> Attaching package: ‘MatrixGenerics’
#> The following objects are masked from ‘package:matrixStats’:
#> 
#>     colAlls, colAnyNAs, colAnys, colAvgsPerRowSet, colCollapse,
#>     colCounts, colCummaxs, colCummins, colCumprods, colCumsums,
#>     colDiffs, colIQRDiffs, colIQRs, colLogSumExps, colMadDiffs,
#>     colMads, colMaxs, colMeans2, colMedians, colMins, colOrderStats,
#>     colProds, colQuantiles, colRanges, colRanks, colSdDiffs, colSds,
#>     colSums2, colTabulates, colVarDiffs, colVars, colWeightedMads,
#>     colWeightedMeans, colWeightedMedians, colWeightedSds,
#>     colWeightedVars, rowAlls, rowAnyNAs, rowAnys, rowAvgsPerColSet,
#>     rowCollapse, rowCounts, rowCummaxs, rowCummins, rowCumprods,
#>     rowCumsums, rowDiffs, rowIQRDiffs, rowIQRs, rowLogSumExps,
#>     rowMadDiffs, rowMads, rowMaxs, rowMeans2, rowMedians, rowMins,
#>     rowOrderStats, rowProds, rowQuantiles, rowRanges, rowRanks,
#>     rowSdDiffs, rowSds, rowSums2, rowTabulates, rowVarDiffs, rowVars,
#>     rowWeightedMads, rowWeightedMeans, rowWeightedMedians,
#>     rowWeightedSds, rowWeightedVars
#> Loading required package: Biobase
#> Welcome to Bioconductor
#> 
#>     Vignettes contain introductory material; view with
#>     'browseVignettes()'. To cite Bioconductor, see
#>     'citation("Biobase")', and for packages 'citation("pkgname")'.
#> 
#> Attaching package: ‘Biobase’
#> The following object is masked from ‘package:MatrixGenerics’:
#> 
#>     rowMedians
#> The following objects are masked from ‘package:matrixStats’:
#> 
#>     anyMissing, rowMedians
#> Loading required package: Biostrings
#> Loading required package: XVector
#> 
#> Attaching package: ‘Biostrings’
#> The following object is masked from ‘package:base’:
#> 
#>     strsplit
#> Loading required package: bumphunter
#> Loading required package: foreach
#> Loading required package: iterators
#> Loading required package: locfit
#> locfit 1.5-9.4 	 2020-03-24
#> Setting options('download.file.method.GEOquery'='auto')
#> Setting options('GEOquery.inmemory.gpl'=FALSE)
#> Loading required package: ChAMPdata
#> Loading required package: DMRcate
#> Warning: replacing previous import 'minfi::getMeth' by 'bsseq::getMeth' when loading 'DMRcate'
#> 
#> Loading required package: Illumina450ProbeVariants.db
#> Loading required package: IlluminaHumanMethylationEPICmanifest
#> Loading required package: DT
#> Loading required package: RPMM
#> Loading required package: cluster
#> No methods found in package 'RSQLite' for request: 'dbListFields' when loading 'lumi'
#> Warning: replacing previous import 'plyr::summarise' by 'plotly::summarise' when loading 'ChAMP'
#> Warning: replacing previous import 'plyr::rename' by 'plotly::rename' when loading 'ChAMP'
#> Warning: replacing previous import 'plyr::arrange' by 'plotly::arrange' when loading 'ChAMP'
#> Warning: replacing previous import 'plyr::mutate' by 'plotly::mutate' when loading 'ChAMP'
#> 
#> Warning: replacing previous import 'plyr::is.discrete' by 'Hmisc::is.discrete' when loading 'ChAMP'
#> Warning: replacing previous import 'plyr::summarize' by 'Hmisc::summarize' when loading 'ChAMP'
#> Warning: replacing previous import 'plotly::subplot' by 'Hmisc::subplot' when loading 'ChAMP'
#> Warning: replacing previous import 'GenomicRanges::sort' by 'globaltest::sort' when loading 'ChAMP'
#> Warning: replacing previous import 'plotly::last_plot' by 'ggplot2::last_plot' when loading 'ChAMP'
#> Warning: replacing previous import 'globaltest::model.matrix' by 'stats::model.matrix' when loading 'ChAMP'
#> Warning: replacing previous import 'globaltest::p.adjust' by 'stats::p.adjust' when loading 'ChAMP'
#> >> Package version 2.20.1 loaded <<
#>        ___ _      _   __  __ ___ 
#>       / __| |_   /_\ |  \/  | _ \
#>      | (__| ' \ / _ \| |\/| |  _/
#>       \___|_||_/_/ \_\_|  |_|_|  
#>       ------------------------------
#>     If you have any question or suggestion about ChAMP, please email to champ450k@gmail.com.
#>     Thank you for citating ChAMP:
#> 
#>     Yuan Tian, Tiffany J Morris, Amy P Webster, Zhen Yang, Stephan Beck, Andrew Feber, Andrew E Teschendorff; ChAMP: updated methylation analysis pipeline for Illumina BeadChips, Bioinformatics, btx513, https://doi.org/10.1093/bioinformatics/btx513
#>      --------------------------
testDir=system.file("extdata",package="ChAMPdata")
myImport <- champ.import(testDir)
#> [===========================]
#> [<<<< ChAMP.IMPORT START >>>>>]
#> -----------------------------
#> 
#> [ Section 1: Read PD Files Start ]
#>   CSV Directory: /data0/Zhongxu/R/x86_64-pc-linux-gnu-library/4.0/ChAMPdata/extdata/lung_test_set.csv
#>   Find CSV Success
#>   Reading CSV File
#> Warning: incomplete final line found on '/data0/Zhongxu/R/x86_64-pc-linux-gnu-library/4.0/ChAMPdata/extdata/lung_test_set.csv'
#>   Replace Sentrix_Position into Array
#>   Replace Sentrix_ID into Slide
#> [ Section 1: Read PD file Done ]
#> 
#> 
#> [ Section 2: Read IDAT files Start ]
#>   Loading:/data0/Zhongxu/R/x86_64-pc-linux-gnu-library/4.0/ChAMPdata/extdata/7990895118_R03C02_Grn.idat ---- (1/8)
#> Warning: truncating string with embedded nuls
#>   Loading:/data0/Zhongxu/R/x86_64-pc-linux-gnu-library/4.0/ChAMPdata/extdata/7990895118_R05C02_Grn.idat ---- (2/8)
#> Warning: truncating string with embedded nuls
#>   Loading:/data0/Zhongxu/R/x86_64-pc-linux-gnu-library/4.0/ChAMPdata/extdata/9247377086_R01C01_Grn.idat ---- (3/8)
#> Warning: truncating string with embedded nuls
#>   Loading:/data0/Zhongxu/R/x86_64-pc-linux-gnu-library/4.0/ChAMPdata/extdata/9247377086_R02C01_Grn.idat ---- (4/8)
#> Warning: truncating string with embedded nuls
#>   Loading:/data0/Zhongxu/R/x86_64-pc-linux-gnu-library/4.0/ChAMPdata/extdata/7766130112_R06C01_Grn.idat ---- (5/8)
#> Warning: truncating string with embedded nuls
#>   Loading:/data0/Zhongxu/R/x86_64-pc-linux-gnu-library/4.0/ChAMPdata/extdata/7766130112_R01C02_Grn.idat ---- (6/8)
#> Warning: truncating string with embedded nuls
#>   Loading:/data0/Zhongxu/R/x86_64-pc-linux-gnu-library/4.0/ChAMPdata/extdata/7990895118_R01C01_Grn.idat ---- (7/8)
#> Warning: truncating string with embedded nuls
#>   Loading:/data0/Zhongxu/R/x86_64-pc-linux-gnu-library/4.0/ChAMPdata/extdata/7990895118_R01C02_Grn.idat ---- (8/8)
#> Warning: truncating string with embedded nuls
#>   Loading:/data0/Zhongxu/R/x86_64-pc-linux-gnu-library/4.0/ChAMPdata/extdata/7990895118_R03C02_Red.idat ---- (1/8)
#> Warning: truncating string with embedded nuls
#>   Loading:/data0/Zhongxu/R/x86_64-pc-linux-gnu-library/4.0/ChAMPdata/extdata/7990895118_R05C02_Red.idat ---- (2/8)
#> Warning: truncating string with embedded nuls
#>   Loading:/data0/Zhongxu/R/x86_64-pc-linux-gnu-library/4.0/ChAMPdata/extdata/9247377086_R01C01_Red.idat ---- (3/8)
#> Warning: truncating string with embedded nuls
#>   Loading:/data0/Zhongxu/R/x86_64-pc-linux-gnu-library/4.0/ChAMPdata/extdata/9247377086_R02C01_Red.idat ---- (4/8)
#> Warning: truncating string with embedded nuls
#>   Loading:/data0/Zhongxu/R/x86_64-pc-linux-gnu-library/4.0/ChAMPdata/extdata/7766130112_R06C01_Red.idat ---- (5/8)
#> Warning: truncating string with embedded nuls
#>   Loading:/data0/Zhongxu/R/x86_64-pc-linux-gnu-library/4.0/ChAMPdata/extdata/7766130112_R01C02_Red.idat ---- (6/8)
#> Warning: truncating string with embedded nuls
#>   Loading:/data0/Zhongxu/R/x86_64-pc-linux-gnu-library/4.0/ChAMPdata/extdata/7990895118_R01C01_Red.idat ---- (7/8)
#> Warning: truncating string with embedded nuls
#>   Loading:/data0/Zhongxu/R/x86_64-pc-linux-gnu-library/4.0/ChAMPdata/extdata/7990895118_R01C02_Red.idat ---- (8/8)
#> Warning: truncating string with embedded nuls
#> 
#>   Extract Mean value for Green and Red Channel Success
#>     Your Red Green Channel contains 622399 probes.
#> [ Section 2: Read IDAT Files Done ]
#> 
#> 
#> [ Section 3: Use Annotation Start ]
#> 
#>   Reading 450K Annotation >>
#> 
#>   Fetching NEGATIVE ControlProbe.
#>     Totally, there are 613 control probes in Annotation.
#>     Your data set contains 613 control probes.
#> 
#>   Generating Meth and UnMeth Matrix
#>     Extracting Meth Matrix...
#>       Totally there are 485512 Meth probes in 450K Annotation.
#>       Your data set contains 485512 Meth probes.
#>     Extracting UnMeth Matrix...
#>       Totally there are 485512 UnMeth probes in 450K Annotation.
#>       Your data set contains 485512 UnMeth probes.
#> 
#>   Generating beta Matrix
#>   Generating M Matrix
#>   Generating intensity Matrix
#>   Calculating Detect P value
#>   Counting Beads
#> [ Section 3: Use Annotation Done ]
#> 
#> [<<<<< ChAMP.IMPORT END >>>>>>]
#> [===========================]
#> [You may want to process champ.filter() next.]

beta.df = myImport$beta
group = myImport$pd$Sample_Group
loonR::ChAMP_QC_Pipeline_Frome_Beta_Value(Sample.beta.df=beta.df, Sample.Group=group)
#> [===========================]
#> [<<<< ChAMP.FILTER START >>>>>]
#> -----------------------------
#> 
#> In New version ChAMP, champ.filter() function has been set to do filtering on the result of champ.import(). You can use champ.import() + champ.filter() to do Data Loading, or set "method" parameter in champ.load() as "ChAMP" to get the same effect.
#> 
#> This function is provided for user need to do filtering on some beta (or M) matrix, which contained most filtering system in champ.load except beadcount. User need to input beta matrix, pd file themselves. If you want to do filterintg on detP matrix and Bead Count, you also need to input a detected P matrix and Bead Count information.
#> 
#> Note that if you want to filter more data matrix, say beta, M, intensity... please make sure they have exactly the same rownames and colnames.
#> 
#> 
#> [ Section 1:  Check Input Start ]
#>   You have inputed beta for Analysis.
#> 
#>   pd file provided, checking if it's in accord with Data Matrix...
#>     pd file check success.
#> 
#>   Parameter filterDetP is TRUE, checking if detP in accord with Data Matrix...
#>     !!! Parameter detP is not found, filterDetP is reset FALSE now.
#> 
#>   Parameter filterBeads is TRUE, checking if beadcount in accord with Data Matrix...
#>     !!! Parameter beadcount is not found, filterBeads is reset FALSE now.
#> 
#>   parameter autoimpute is TRUE. Checking if the conditions are fulfilled...
#>     !!! ProbeCutoff is 0, which means you have no needs to do imputation. autoimpute has been reset FALSE.
#> 
#>   Checking Finished :filterMultiHit,filterSNPs,filterNoCG,filterXY would be done on beta.
#> [ Section 1: Check Input Done ]
#> 
#> 
#> [ Section 2: Filtering Start >>
#> 
#>   Filtering NoCG Start
#>     Only Keep CpGs, removing 3091 probes from the analysis.
#> 
#>   Filtering SNPs Start
#>     Using general 450K SNP list for filtering.
#>     Filtering probes with SNPs as identified in Zhou's Nucleic Acids Research Paper 2016.
#>     Removing 59901 probes from the analysis.
#> 
#>   Filtering MultiHit Start
#>     Filtering probes that align to multiple locations as identified in Nordlund et al
#>     Removing 11 probes from the analysis.
#> 
#>   Filtering XY Start
#>     Filtering probes located on X,Y chromosome, removing 10028 probes from the analysis.
#> 
#>   Updating PD file
#>     filterDetP parameter is FALSE, so no Sample Would be removed.
#> [ Section 2: Filtering Done ]
#> 
#>  All filterings are Done, now you have 412481 probes and 8 samples.
#> [<<<<< ChAMP.FILTER END >>>>>>]
#> [===========================]
#> [You may want to process champ.QC() next.]
#> [===========================]
#> [<<< ChAMP.IMPUTE START >>>>]
#> -----------------------------
#> <method>:Combine. (Suitable for Large Data Set)
#> Error in apply(beta, 2, function(x) length(which(is.na(x)))/length(x)) object 'myLoad' not found