Plotting the Net benefit function. Coule be multiple curves

Plotting the Net benefit function. Coule be multiple curves

# S3 method for ntbft
plot(
  nb,
  nolines = 2:dim(nb)[2],
  nobands = NULL,
  ymin = -0.1,
  ymax = max(nb[, c(nolines, nobands)], na.rm = T),
  legpos = c(0.9, 0.8),
  palette = "aaas"
)

Arguments

nb
Object from

nolines
the number of the columns of nb which should be plotted using lines. The default is to plot all columns (except the first one containing the threshold).

nobands
the number of the columns of nb which should be plotted using bands (useful to plot confidence intervals). The default is to plot no bands

legpos
a vector of two coordinates indicating where the legend should be in the graph

palette

Value

Examples

p = loonR::decisionCurveAnalysisSimple(label, risk)
#> Error in unique.default(response) unique() applies only to vectors
p$Plot
#> Error in eval(expr, envir, enclos) object 'p' not found
v.netbenefit = p$NetBenefit
#> Error in eval(expr, envir, enclos) object 'p' not found

# similar to getr.netbenefit

nb <- data.frame(threshold = v.netbenefit$threshold,
                 All = v.netbenefit$All,
                 None = v.netbenefit$None,
                 `Transcriptomic panel` = v.netbenefit$Model,
                 `Risk-stratification model` = r.netbenefit$Model,
                 stringsAsFactors = F,
                 check.names = F)
#> Error in data.frame(threshold = v.netbenefit$threshold, All = v.netbenefit$All,     None = v.netbenefit$None, `Transcriptomic panel` = v.netbenefit$Model,     `Risk-stratification model` = r.netbenefit$Model, stringsAsFactors = F,     check.names = F) object 'v.netbenefit' not found

p = loonR:::plot.ntbft(nb, 2:5)
#> Error in loonR:::plot.ntbft(nb, 2:5) object 'nb' not found
p
#> Error in eval(expr, envir, enclos) object 'p' not found